feat(test): v0.49.0 — test infrastructure hardening & scheduler decomposition
TEST-10-01: Replace tokio::time::sleep with pg_stat_activity polling in
  concurrent E2E tests (wait_for_active_query helper).

TEST-10-02: Add unit tests to template_cache.rs, cdc/polling.rs, cdc/rebuild.rs
  covering hash key derivation, trigger naming, CDC mode classification,
  and replica identity sufficiency.

TEST-10-03: Add fuzz targets merge_sql_fuzz (FUZZ-7) and row_id_fuzz (FUZZ-8)
  with public wrapper module src/fuzz_pub.rs.

TEST-10-04: Add test_ddl_during_concurrent_refresh to e2e_concurrent_tests.rs.

CI-10-02: Expand e2e-smoke filter to match join/aggregate/window/subquery tests.

CI-10-03: Add just fuzz-all recipe; document fuzz targets in CONTRIBUTING.md.

CQ-10-01: Extract scheduler submodules:
  - src/scheduler/dispatch.rs  (parallel dispatch, worker spawn/reap)
  - src/scheduler/scheduler_loop.rs  (BGW registration, main loops)
  - src/scheduler/w  - src/scheduler/w  - src/scheduler/w  - src/scheduler/w  - src/scheduler/w  - src/scheduler/w  - src/scheduler/w  - src/scheduler/w  - src.49.0, update CHANGELOG.md, ROADMAP.md,
  roadmap/v0.49.0.md-full.md.
