+++ .github/workflows/ci.yml
@@ -284,7 +284,7 @@ jobs:
         run: |
           cargo nextest run --profile ci \
             --test 'e2e_*' \
-            -E "test(smoke) | test(basic) | test(create_stream_table) | test(health_check)" \
+            -E "test(smoke) | test(basic) | test(create_stream_table) | test(health_check) | test(~join) | test(~aggregate) | test(~window) | test(~subquery)" \
             --test-threads 4
 
   # ── E2E tests (full extension in Docker, Linux only) ────────────────────
