+++ Cargo.toml
@@ -1,6 +1,6 @@
 [package]
 name = "pg_trickle"
-version = "0.48.0"
+version = "0.49.0"
 edition = "2024"
 description = "Streaming Stream Tables for PostgreSQL 18 with differential view maintenance"
 license = "Apache-2.0"
