+++ src/scheduler/mod.rs
@@ -32,37 +32,44 @@
 //! - **Error classification**: only retryable errors trigger retry; user/schema
 //!   errors fail immediately
 
-use pgrx::bgworkers::*;
 use pgrx::prelude::*;
 
 use std::cell::RefCell;
-use std::collections::{HashMap, HashSet, VecDeque};
+use std::collections::HashMap;
 use std::panic::AssertUnwindSafe;
 
-use crate::catalog::{JobStatus, SchedulerJob};
+use crate::catalog::SchedulerJob;
 use crate::catalog::{RefreshRecord, StreamTableMeta};
 use crate::cdc;
 use crate::config;
-use crate::dag::{
-    DiamondConsistency, DiamondSchedulePolicy, ExecutionUnit, ExecutionUnitDag, ExecutionUnitId,
-    NodeId, RefreshMode, Scc, StDag, StStatus,
-};
+use crate::dag::{DiamondSchedulePolicy, NodeId, RefreshMode, Scc, StDag, StStatus};
 use crate::error::{RetryPolicy, RetryState};
 use crate::monitor;
 use crate::refresh::{self, RefreshAction};
 use crate::shmem;
 use crate::version;
-use crate::wal_decoder;
 
 pub mod citus;
 pub mod cost;
+pub mod dispatch;
 pub mod pool;
+pub mod scheduler_loop;
 pub mod tier;
+pub mod watermark;
+
+// CQ-10-01: Re-export public items from decomposed scheduler submodules.
+pub use dispatch::{reconcile_parallel_state, spawn_refresh_worker};
+pub use scheduler_loop::{
+    pg_trickle_launcher_main, pg_trickle_scheduler_main, register_launcher_worker,
+    register_scheduler_worker,
+};
 
 use citus::drive_distributed_cdc;
 pub use cost::compute_per_db_quota;
 pub use cost::{compute_per_db_quota_with_lag, lag_aware_quota_boost};
+use dispatch::extract_panic_message;
 pub use tier::RefreshTier;
+use watermark::compute_worker_tick_watermark;
 
 // ── SCAL-1 (v0.25.0): Per-backend catalog snapshot cache ─────────────────
 //
@@ -218,822 +225,6 @@ impl Drop for SubTransaction {
             }
         }
     }
-}
-
-/// Register the launcher background worker.
-///
-/// Called from `_PG_init()` when loaded via `shared_preload_libraries`.
-/// The launcher discovers all databases on the server and spawns a separate
-/// per-database scheduler worker for each one that has pg_trickle installed.
-pub fn register_launcher_worker() {
-    BackgroundWorkerBuilder::new("pg_trickle launcher")
-        .set_function("pg_trickle_launcher_main")
-        .set_library("pg_trickle")
-        .enable_spi_access()
-        .set_start_time(BgWorkerStartTime::RecoveryFinished)
-        .set_restart_time(Some(std::time::Duration::from_secs(5)))
-        .load();
-}
-
-/// Main entry point for the launcher background worker.
-///
-/// Runs once per server. Connects to `postgres` (always present) to query
-/// `pg_database` and `pg_stat_activity`, then dynamically spawns a scheduler
-/// worker for every database where pg_trickle is installed.
-///
-/// # Safety
-/// Called directly by PostgreSQL as a background worker entry point.
-#[pg_guard]
-#[unsafe(no_mangle)]
-pub extern "C-unwind" fn pg_trickle_launcher_main(_arg: pg_sys::Datum) {
-    BackgroundWorker::attach_signal_handlers(SignalWakeFlags::SIGHUP | SignalWakeFlags::SIGTERM);
-    // Connect to `postgres` — always available; used only for system catalog queries.
-    BackgroundWorker::connect_worker_to_spi(Some("postgres"), None);
-
-    // Check for replica — launcher cannot spawn workers on standbys.
-    let is_replica = BackgroundWorker::transaction(AssertUnwindSafe(|| -> bool {
-        Spi::get_one::<bool>("SELECT pg_is_in_recovery()")
-            .unwrap_or(Some(false))
-            .unwrap_or(false)
-    }));
-    if is_replica {
-        log!("pg_trickle launcher: running on read replica, sleeping until promotion");
-        loop {
-            let ok = BackgroundWorker::wait_latch(Some(std::time::Duration::from_secs(30)));
-            if !ok {
-                return;
-            }
-            let still = BackgroundWorker::transaction(AssertUnwindSafe(|| -> bool {
-                Spi::get_one::<bool>("SELECT pg_is_in_recovery()")
-                    .unwrap_or(Some(false))
-                    .unwrap_or(false)
-            }));
-            if !still {
-                break;
-            }
-        }
-    }
-
-    log!("pg_trickle launcher started");
-
-    // Track the DAG rebuild signal so we can detect CREATE EXTENSION in any
-    // database and immediately re-probe, rather than waiting for skip_ttl.
-    let mut last_dag_version = shmem::current_dag_version();
-
-    // last_attempt[db] = when we last tried to spawn a worker for that DB.
-    // Used to avoid hammering databases where pg_trickle is not installed.
-    let mut last_attempt: HashMap<String, std::time::Instant> = HashMap::new();
-    // DBs where we have seen a scheduler successfully start at least once this
-    // session. Used to distinguish "never had pg_trickle" (long backoff) from
-    // "had a running scheduler that crashed" (short backoff).
-    let mut had_scheduler: HashSet<String> = HashSet::new();
-    // How long to wait before re-probing a DB that has never had pg_trickle.
-    let skip_ttl = std::time::Duration::from_secs(300);
-    // How long to wait before respawning a scheduler that crashed / exited
-    // on a DB where pg_trickle was previously confirmed running.  Kept short
-    // so DROP EXTENSION + CREATE EXTENSION doesn't stall for 5 minutes.
-    let retry_ttl = std::time::Duration::from_secs(15);
-
-    // STAB-3 (v0.30.0): Track last time we ran the age-based template-cache purge.
-    // The purge runs at most once per hour per launcher tick to avoid per-tick
-    // catalog load. It is not per-database (the launcher connects to `postgres`).
-    let last_cache_purge = std::time::Instant::now()
-        .checked_sub(std::time::Duration::from_secs(7200))
-        .unwrap_or_else(std::time::Instant::now);
-
-    loop {
-        // ── Collect all connectable, non-template databases  ──────────────
-        let databases: Vec<String> = BackgroundWorker::transaction(AssertUnwindSafe(|| {
-            Spi::connect(|client| -> Vec<String> {
-                match client.select(
-                    "SELECT datname::text FROM pg_database \
-                     WHERE NOT datistemplate AND datallowconn",
-                    None,
-                    &[],
-                ) {
-                    Ok(result) => {
-                        let mut out = Vec::new();
-                        for row in result {
-                            if let Some(db) = row.get_by_name::<String, _>("datname").ok().flatten()
-                            {
-                                out.push(db);
-                            }
-                        }
-                        out
-                    }
-                    Err(_) => vec![],
-                }
-            })
-        }));
-
-        // ── Databases that already have a running scheduler worker  ────────
-        let active: HashSet<String> = BackgroundWorker::transaction(AssertUnwindSafe(|| {
-            Spi::connect(|client| -> HashSet<String> {
-                match client.select(
-                    "SELECT datname::text \
-                     FROM pg_stat_activity \
-                     WHERE backend_type = 'pg_trickle scheduler' \
-                       AND datname IS NOT NULL",
-                    None,
-                    &[],
-                ) {
-                    Ok(result) => {
-                        let mut out = HashSet::new();
-                        for row in result {
-                            if let Some(db) = row.get_by_name::<String, _>("datname").ok().flatten()
-                            {
-                                out.insert(db);
-                            }
-                        }
-                        out
-                    }
-                    Err(_) => HashSet::new(),
-                }
-            })
-        }));
-
-        // If any backend bumped the DAG signal (create_st, alter, drop,
-        // CREATE EXTENSION finalize, manual rescan nudge, etc.) since our last
-        // loop, clear the skip-cache so we re-probe promptly.
-        //
-        // This is intentionally unconditional. A CREATE EXTENSION in a newly
-        // created database can race with the launcher's first probe of that
-        // database: the launcher may spawn a scheduler before the extension
-        // transaction commits, observe "not installed", and cache a fresh
-        // `last_attempt` entry. If we preserve that fresh entry on the very DAG
-        // signal emitted by CREATE EXTENSION, the database remains classified as
-        // a long-backoff `skip_ttl` database and the scheduler may not appear
-        // for minutes.
-        //
-        // DDL-driven DAG bumps are relatively rare and semantically important,
-        // so favor immediate correctness over preserving the old skip-cache on
-        // those events. Non-DDL steady-state operation still uses the normal
-        // skip_ttl / retry_ttl back-off between signals.
-        let dag_version = shmem::current_dag_version();
-        if dag_version != last_dag_version {
-            last_dag_version = dag_version;
-            last_attempt.clear();
-        }
-
-        for db in &databases {
-            if active.contains(db) {
-                // Worker healthy — record that this DB has had a running
-                // scheduler so we can use the short retry_ttl if it crashes.
-                had_scheduler.insert(db.clone());
-                last_attempt.remove(db);
-                continue;
-            }
-
-            // Choose the right backoff: short for DBs that previously ran a
-            // scheduler (crash / DROP EXTENSION scenario), long for DBs that
-            // have never had pg_trickle installed.
-            let ttl = if had_scheduler.contains(db) {
-                retry_ttl
-            } else {
-                skip_ttl
-            };
-
-            // Should we (re)try this database?
-            let retry = last_attempt
-                .get(db)
-                .map(|t| t.elapsed() >= ttl)
-                .unwrap_or(true);
-            if !retry {
-                continue;
-            }
-
-            last_attempt.insert(db.clone(), std::time::Instant::now());
-
-            match BackgroundWorkerBuilder::new("pg_trickle scheduler")
-                .set_function("pg_trickle_scheduler_main")
-                .set_library("pg_trickle")
-                .enable_spi_access()
-                // Pass the database name via bgw_extra (max 128 bytes).
-                .set_extra(db.as_str())
-                // No restart_time — the launcher itself handles respawning.
-                .set_restart_time(None)
-                .load_dynamic()
-            {
-                Ok(_) => {
-                    log!(
-                        "pg_trickle launcher: spawned scheduler for database '{}'",
-                        db
-                    );
-                }
-                Err(_) => {
-                    warning!(
-                        "pg_trickle launcher: could not spawn scheduler for database '{}'",
-                        db
-                    );
-                }
-            }
-        }
-
-        // STAB-3 (v0.30.0): Age-based purge of the L2 catalog template cache.
-        // Run at most once per hour. The launcher connects to `postgres` which
-        // does not have pg_trickle installed, so we skip the purge here — the
-        // per-database scheduler workers run the purge in their own DB context.
-        let _ = last_cache_purge; // suppress unused-variable warning
-
-        // Wake every 10 s or on SIGHUP/SIGTERM.
-        let should_continue =
-            BackgroundWorker::wait_latch(Some(std::time::Duration::from_secs(10)));
-
-        unsafe {
-            if pg_sys::ConfigReloadPending != 0 {
-                pg_sys::ConfigReloadPending = 0;
-                pg_sys::ProcessConfigFile(pg_sys::GucContext::PGC_SIGHUP);
-                pgrx::info!(
-                    "pg_trickle scheduler: SIGHUP processed, allow_circular is now {}",
-                    config::pg_trickle_allow_circular()
-                );
-            }
-        }
-
-        if !should_continue {
-            log!("pg_trickle launcher shutting down");
-            break;
-        }
-    }
-}
-
-/// Register the scheduler background worker (single-database, legacy).
-///
-/// Kept for reference — `_PG_init()` now registers the launcher instead.
-/// The launcher dynamically spawns instances of this worker, one per database.
-pub fn register_scheduler_worker() {
-    BackgroundWorkerBuilder::new("pg_trickle scheduler")
-        .set_function("pg_trickle_scheduler_main")
-        .set_library("pg_trickle")
-        .enable_spi_access()
-        .set_start_time(BgWorkerStartTime::RecoveryFinished)
-        .set_restart_time(Some(std::time::Duration::from_secs(5)))
-        .load();
-}
-
-// ── Dynamic Refresh Worker (Phase 3) ──────────────────────────────────────
-
-/// Spawn a dynamic refresh worker for a specific job.
-///
-/// The coordinator calls this after enqueuing a job and acquiring a worker
-/// token from shared memory. The database name and job_id are packed into
-/// `bgw_extra` as `"<db_name>|<job_id>"`.
-///
-/// **Why `|` and not `\0`?** pgrx's `BackgroundWorker::get_extra()` uses
-/// `CStr::from_ptr` internally, which treats the first null byte as the
-/// string terminator. A null-byte separator would therefore cause the
-/// job_id portion to be silently dropped when the worker reads the extra.
-/// The pipe character `|` is not a valid PostgreSQL identifier character
-/// (without quoting) and is safe to use as a delimiter here.
-///
-/// Returns `Ok(())` on successful spawn, `Err(...)` if the dynamic worker
-/// could not be registered.
-pub fn spawn_refresh_worker(
-    db_name: &str,
-    job_id: i64,
-) -> Result<(), crate::error::PgTrickleError> {
-    // Pack db_name + job_id into bgw_extra (max 128 bytes).
-    // Use '|' as separator — see doc comment above for why not '\0'.
-    let extra = format!("{db_name}|{job_id}");
-    if extra.len() > 128 {
-        return Err(crate::error::PgTrickleError::InternalError(format!(
-            "bgw_extra too long ({} bytes) for db='{}' job_id={}",
-            extra.len(),
-            db_name,
-            job_id
-        )));
-    }
-
-    BackgroundWorkerBuilder::new("pg_trickle refresh worker")
-        .set_function("pg_trickle_refresh_worker_main")
-        .set_library("pg_trickle")
-        .enable_spi_access()
-        .set_extra(&extra)
-        .set_restart_time(None) // no auto-restart — coordinator handles retries
-        .load_dynamic()
-        .map_err(|_| {
-            crate::error::PgTrickleError::InternalError(
-                "Failed to register dynamic refresh worker".into(),
-            )
-        })?;
-
-    Ok(())
-}
-
-/// Main entry point for a dynamic refresh worker (Phase 3).
-///
-/// Each refresh worker:
-/// 1. Parses `db_name` and `job_id` from `bgw_extra`.
-/// 2. Connects to the target database.
-/// 3. Claims the job (QUEUED → RUNNING).
-/// 4. Executes the execution unit (singleton for now, composite later).
-/// 5. Persists the outcome to the job table.
-/// 6. Releases the worker token from shared memory.
-///
-/// # Safety
-/// Called directly by PostgreSQL as a background worker entry point.
-#[pg_guard]
-#[unsafe(no_mangle)]
-pub extern "C-unwind" fn pg_trickle_refresh_worker_main(_arg: pg_sys::Datum) {
-    BackgroundWorker::attach_signal_handlers(SignalWakeFlags::SIGHUP | SignalWakeFlags::SIGTERM);
-
-    // Parse bgw_extra: "db_name\0job_id"
-    let extra = BackgroundWorker::get_extra();
-    let (db_name, job_id) = match parse_worker_extra(extra) {
-        Some(pair) => pair,
-        None => {
-            warning!(
-                "pg_trickle refresh worker: malformed bgw_extra '{}', exiting",
-                extra
-            );
-            shmem::release_worker_token();
-            return;
-        }
-    };
-
-    BackgroundWorker::connect_worker_to_spi(Some(&db_name), None);
-
-    // Get our own PID for logging and job claiming.
-    let my_pid = BackgroundWorker::transaction(AssertUnwindSafe(|| -> i32 {
-        match Spi::get_one::<i32>("SELECT pg_backend_pid()") {
-            Ok(Some(pid)) => pid,
-            Ok(None) => {
-                pgrx::warning!(
-                    "pg_trickle scheduler: pg_backend_pid() returned NULL (db='{}')",
-                    db_name
-                );
-                0
-            }
-            Err(e) => {
-                pgrx::warning!(
-                    "pg_trickle scheduler: could not fetch backend pid (db='{}'): {}",
-                    db_name,
-                    e
-                );
-                0
-            }
-        }
-    }));
-
-    log!(
-        "pg_trickle refresh worker: started (db='{}', job_id={}, pid={})",
-        db_name,
-        job_id,
-        my_pid,
-    );
-
-    // Claim the job: QUEUED → RUNNING
-    let claimed = BackgroundWorker::transaction(AssertUnwindSafe(|| -> bool {
-        match SchedulerJob::claim(job_id, my_pid) {
-            Ok(v) => v,
-            Err(e) => {
-                pgrx::warning!(
-                    "pg_trickle scheduler: failed to claim job {} (db='{}', pid={}): {}",
-                    job_id,
-                    db_name,
-                    my_pid,
-                    e
-                );
-                false
-            }
-        }
-    }));
-
-    if !claimed {
-        log!(
-            "pg_trickle refresh worker: job {} already claimed or cancelled, exiting",
-            job_id
-        );
-        shmem::release_worker_token();
-        return;
-    }
-
-    // Load the job details
-    let job = BackgroundWorker::transaction(AssertUnwindSafe(|| -> Option<SchedulerJob> {
-        SchedulerJob::get_by_id(job_id).ok().flatten()
-    }));
-
-    let job = match job {
-        Some(j) => j,
-        None => {
-            log!(
-                "pg_trickle refresh worker: job {} not found after claim, exiting",
-                job_id
-            );
-            BackgroundWorker::transaction(AssertUnwindSafe(|| {
-                let _ = SchedulerJob::cancel(job_id, "Job row disappeared after claim");
-            }));
-            shmem::release_worker_token();
-            return;
-        }
-    };
-
-    // Validate: DAG version hasn't become obsolete
-    let current_dag_version = shmem::current_dag_version();
-    if (current_dag_version as i64) > job.dag_version + 1 {
-        log!(
-            "pg_trickle refresh worker: job {} dag_version={} is stale (current={}), cancelling",
-            job_id,
-            job.dag_version,
-            current_dag_version,
-        );
-        BackgroundWorker::transaction(AssertUnwindSafe(|| {
-            let _ = SchedulerJob::cancel(job_id, "DAG version obsolete");
-        }));
-        shmem::release_worker_token();
-        return;
-    }
-
-    // Execute the unit.
-    //
-    // ERR-1d: Wrap in catch_unwind so that PostgreSQL ERRORs (which pgrx
-    // converts to Rust panics via longjmp) don't kill the worker before we
-    // can record the failure. When the refresh triggers a PG ERROR (e.g.
-    // "column X does not exist"), the BackgroundWorker::transaction's
-    // PgTryBuilder rethrows the caught panic, aborting the transaction.
-    // Without catch_unwind, the panic propagates to #[pg_guard] and the
-    // worker exits — leaving the job stuck in RUNNING and the ST never
-    // transitioning to ERROR.
-    let outcome = std::panic::catch_unwind(AssertUnwindSafe(|| {
-        BackgroundWorker::transaction(AssertUnwindSafe(|| -> RefreshOutcome {
-            match job.unit_kind.as_str() {
-                "singleton" => execute_worker_singleton(&job),
-                "atomic_group" => execute_worker_atomic_group(&job, false),
-                "repeatable_read_group" => execute_worker_atomic_group(&job, true),
-                "immediate_closure" => execute_worker_immediate_closure(&job),
-                "cyclic_scc" => execute_worker_cyclic_scc(&job),
-                "fused_chain" => execute_worker_fused_chain(&job),
-                _ => {
-                    warning!(
-                        "pg_trickle refresh worker: unknown unit_kind '{}' for job {}",
-                        job.unit_kind,
-                        job_id,
-                    );
-                    RefreshOutcome::PermanentFailure
-                }
-            }
-        }))
-    }));
-
-    let outcome = match outcome {
-        Ok(o) => (o, None),
-        Err(panic_payload) => {
-            // ERR-1d: The refresh transaction panicked (PG ERROR). Extract
-            // the error message from the panic payload and treat this as a
-            // permanent failure. The original transaction was rolled back by
-            // PostgreSQL, so we record the failure in a fresh transaction.
-            let error_msg = extract_panic_message(&panic_payload);
-
-            log!(
-                "pg_trickle refresh worker: job {} panicked (PG ERROR): {}",
-                job_id,
-                error_msg,
-            );
-
-            // ERR-1d fix: After a PG ERROR inside BackgroundWorker::transaction,
-            // the PostgreSQL transaction state is left in STARTED/ABORT state
-            // because CommitTransactionCommand was never called.  We must abort
-            // the orphaned transaction before starting a new one, otherwise
-            // StartTransactionCommand raises "unexpected state STARTED".
-            // SAFETY: AbortCurrentTransaction properly cleans up the aborted
-            // transaction and returns PostgreSQL to idle state.
-            unsafe {
-                pg_sys::AbortCurrentTransaction();
-            }
-
-            // Determine if this is a retryable or permanent error.
-            // O39-6: use SQLSTATE-first classifier when use_sqlstate_classification=true.
-            let is_retryable = crate::error::classify_error_for_retry(&error_msg);
-
-            // Set error state on member STs in a fresh transaction.
-            if !is_retryable {
-                BackgroundWorker::transaction(AssertUnwindSafe(|| {
-                    for &pgt_id in &job.member_pgt_ids {
-                        let _ = StreamTableMeta::set_error_state(pgt_id, &error_msg);
-                    }
-                }));
-            }
-
-            let outcome = if is_retryable {
-                RefreshOutcome::RetryableFailure
-            } else {
-                RefreshOutcome::PermanentFailure
-            };
-            (outcome, Some(error_msg))
-        }
-    };
-
-    let (outcome, panic_error_msg) = outcome;
-
-    // Persist outcome to the job table
-    let (status, retryable) = match outcome {
-        RefreshOutcome::Success => (JobStatus::Succeeded, None),
-        RefreshOutcome::RetryableFailure => (JobStatus::RetryableFailed, Some(true)),
-        RefreshOutcome::PermanentFailure => (JobStatus::PermanentFailed, Some(false)),
-    };
-
-    BackgroundWorker::transaction(AssertUnwindSafe(|| {
-        let _ = SchedulerJob::complete(job_id, status, panic_error_msg.as_deref(), retryable);
-    }));
-
-    log!(
-        "pg_trickle refresh worker: job {} finished with {} (db='{}')",
-        job_id,
-        status,
-        db_name,
-    );
-
-    // Release the cluster-wide worker token
-    shmem::release_worker_token();
-}
-
-/// ERR-1d: Extract a human-readable error message from a caught panic payload.
-///
-/// Panics from PostgreSQL ERRORs are represented as `CaughtError` in pgrx.
-/// We try to downcast to known types; if that fails, we produce a generic message.
-fn extract_panic_message(payload: &Box<dyn std::any::Any + Send>) -> String {
-    use pgrx::pg_sys::panic::CaughtError;
-
-    if let Some(caught) = payload.downcast_ref::<CaughtError>() {
-        return match caught {
-            CaughtError::PostgresError(ereport)
-            | CaughtError::ErrorReport(ereport)
-            | CaughtError::RustPanic { ereport, .. } => {
-                // Include full debug output to capture DETAIL, HINT, etc.
-                format!("{:?}", ereport)
-            }
-        };
-    }
-    if let Some(msg) = payload.downcast_ref::<&str>() {
-        return msg.to_string();
-    }
-    if let Some(msg) = payload.downcast_ref::<String>() {
-        return msg.clone();
-    }
-    "unknown error (panic payload could not be decoded)".to_string()
-}
-
-/// Parse the worker's bgw_extra string: "db_name\0job_id".
-fn parse_worker_extra(extra: &str) -> Option<(String, i64)> {
-    // Format is "<db_name>|<job_id>" — see spawn_refresh_worker for why '|'
-    // is used instead of '\0' (pgrx's get_extra() truncates at null bytes).
-    let parts: Vec<&str> = extra.splitn(2, '|').collect();
-    if parts.len() != 2 {
-        return None;
-    }
-    let db_name = parts[0].to_string();
-    let job_id = parts[1].parse::<i64>().ok()?;
-    if db_name.is_empty() || job_id <= 0 {
-        return None;
-    }
-    Some((db_name, job_id))
-}
-
-// ── #536: Frontier holdback tick watermark helpers ─────────────────────────
-
-/// Unix-epoch timestamp of the last holdback-active WARNING, used to
-/// rate-limit warnings to at most one per minute.
-static LAST_HOLDBACK_WARN_SECS: std::sync::atomic::AtomicU64 = std::sync::atomic::AtomicU64::new(0);
-
-/// Compute the tick watermark for the **coordinator** (main scheduler loop).
-///
-/// Applies the `frontier_holdback_mode` GUC logic:
-/// - `"none"` / watermark disabled: use raw `pg_current_wal_lsn()`.
-/// - `"xmin"`: probe `pg_stat_activity` + `pg_prepared_xacts` and hold back
-///   if a long-running transaction would cause data loss.
-/// - `"lsn:<N>"`: hold back by exactly N bytes.
-///
-/// Side effects (when holdback fires):
-/// - Updates `shmem::last_tick_oldest_xmin` for the next tick.
-/// - Updates `shmem::last_tick_safe_lsn_u64` for dynamic workers.
-/// - Updates the holdback gauge metrics.
-/// - Emits a WARNING when holdback age exceeds the warn threshold.
-///
-/// # Arguments
-/// - `prev_watermark_lsn`: the safe LSN from the previous tick, if any.
-///
-/// # Returns
-/// `(tick_watermark, current_oldest_xmin, oldest_txn_age_secs)`
-fn compute_coordinator_tick_watermark(
-    prev_watermark_lsn: Option<&str>,
-) -> (Option<String>, u64, u64) {
-    if !config::pg_trickle_tick_watermark_enabled() {
-        return (None, 0, 0);
-    }
-
-    let mode = config::pg_trickle_frontier_holdback_mode();
-
-    match mode {
-        config::FrontierHoldbackMode::None => {
-            let lsn = match Spi::get_one::<String>("SELECT pg_current_wal_lsn()::text") {
-                Ok(v) => v,
-                Err(e) => {
-                    pgrx::warning!(
-                        "pg_trickle scheduler: failed to fetch pg_current_wal_lsn (mode=None): {}",
-                        e
-                    );
-                    None
-                }
-            };
-            // Store raw write LSN for workers.
-            if let Some(ref l) = lsn {
-                shmem::set_last_tick_safe_lsn(version::lsn_to_u64(l));
-            }
-            shmem::update_holdback_metrics(0, 0);
-            (lsn, 0, 0)
-        }
-
-        config::FrontierHoldbackMode::Xmin | config::FrontierHoldbackMode::InvalidLsn => {
-            // Skip the probe when CDC mode is WAL -- commit-LSN ordering
-            // is already safe in logical-replication mode.
-            if config::pg_trickle_cdc_mode() == "wal" {
-                let lsn = match Spi::get_one::<String>("SELECT pg_current_wal_lsn()::text") {
-                    Ok(v) => v,
-                    Err(e) => {
-                        pgrx::warning!(
-                            "pg_trickle scheduler: failed to fetch pg_current_wal_lsn (mode=wal): {}",
-                            e
-                        );
-                        None
-                    }
-                };
-                if let Some(ref l) = lsn {
-                    shmem::set_last_tick_safe_lsn(version::lsn_to_u64(l));
-                }
-                shmem::update_holdback_metrics(0, 0);
-                return (lsn, 0, 0);
-            }
-
-            let prev_oldest_xmin = shmem::last_tick_oldest_xmin();
-
-            match cdc::compute_safe_upper_bound(prev_watermark_lsn, prev_oldest_xmin) {
-                Ok((safe_lsn, write_lsn, current_oldest_xmin, age_secs)) => {
-                    // Persist for next tick and for dynamic workers under a
-                    // single lock so workers never see xmin/LSN out of sync.
-                    let safe_u64 = version::lsn_to_u64(&safe_lsn);
-                    shmem::set_last_tick_holdback_state(current_oldest_xmin, safe_u64);
-
-                    // Update holdback gauge metrics.
-                    let write_u64 = version::lsn_to_u64(&write_lsn);
-                    let holdback_bytes = write_u64.saturating_sub(safe_u64);
-                    shmem::update_holdback_metrics(holdback_bytes, age_secs);
-
-                    // Warn when holdback has been active longer than the threshold.
-                    if holdback_bytes > 0 {
-                        emit_holdback_warning_if_needed(age_secs);
-                    }
-
-                    (Some(safe_lsn), current_oldest_xmin, age_secs)
-                }
-                Err(e) => {
-                    // On probe failure, hold at the previous watermark (if known)
-                    // rather than advancing to the raw write LSN.  Advancing on
-                    // failure is the exact unsafe behaviour the holdback is meant
-                    // to prevent — the probe may have failed precisely because a
-                    // long-running transaction exists.
-                    warning!(
-                        "pg_trickle: holdback probe failed ({}); holding at previous watermark",
-                        e
-                    );
-                    let safe_lsn = match prev_watermark_lsn {
-                        Some(prev) => {
-                            // Re-use last known-safe watermark.
-                            let u = version::lsn_to_u64(prev);
-                            shmem::set_last_tick_safe_lsn(u);
-                            Some(prev.to_string())
-                        }
-                        None => {
-                            // First tick — no previous watermark; fall back to
-                            // write LSN to avoid stalling forever on startup.
-                            let lsn = Spi::get_one::<String>("SELECT pg_current_wal_lsn()::text")
-                                .unwrap_or(None);
-                            if let Some(ref l) = lsn {
-                                shmem::set_last_tick_safe_lsn(version::lsn_to_u64(l));
-                            }
-                            lsn
-                        }
-                    };
-                    shmem::update_holdback_metrics(0, 0);
-                    (safe_lsn, 0, 0)
-                }
-            }
-        }
-
-        config::FrontierHoldbackMode::LsnBytes(offset_bytes) => {
-            let write_lsn_str = Spi::get_one::<String>("SELECT pg_current_wal_lsn()::text")
-                .unwrap_or(None)
-                .unwrap_or_else(|| "0/0".to_string());
-            let write_u64 = version::lsn_to_u64(&write_lsn_str);
-            let safe_u64 = write_u64.saturating_sub(offset_bytes);
-            let safe_lsn = version::u64_to_lsn(safe_u64);
-            shmem::set_last_tick_safe_lsn(safe_u64);
-            shmem::update_holdback_metrics(offset_bytes.min(write_u64), 0);
-            (Some(safe_lsn), 0, 0)
-        }
-    }
-}
-
-/// Compute the tick watermark for a **dynamic refresh worker**.
-///
-/// Dynamic workers run after the coordinator and do not have access to
-/// the previous tick's `prev_watermark_lsn`. They read the coordinator-
-/// computed safe watermark from shared memory and cap it with the current
-/// write LSN (in case the worker starts significantly after the tick).
-///
-/// When holdback is disabled or shmem is unavailable, falls back to
-/// `pg_current_wal_lsn()`.
-fn compute_worker_tick_watermark() -> Option<String> {
-    if !config::pg_trickle_tick_watermark_enabled() {
-        return None;
-    }
-
-    let mode = config::pg_trickle_frontier_holdback_mode();
-
-    match mode {
-        config::FrontierHoldbackMode::None => {
-            match Spi::get_one::<String>("SELECT pg_current_wal_lsn()::text") {
-                Ok(v) => v,
-                Err(e) => {
-                    pgrx::warning!(
-                        "pg_trickle scheduler: failed to fetch pg_current_wal_lsn (worker/None): {}",
-                        e
-                    );
-                    None
-                }
-            }
-        }
-
-        config::FrontierHoldbackMode::Xmin
-        | config::FrontierHoldbackMode::LsnBytes(_)
-        | config::FrontierHoldbackMode::InvalidLsn => {
-            // Read the safe watermark the coordinator stored in shmem.
-            let safe_lsn_u64 = shmem::last_tick_safe_lsn_u64();
-
-            if safe_lsn_u64 == 0 {
-                // No coordinator value yet — fall back to raw write LSN.
-                return match Spi::get_one::<String>("SELECT pg_current_wal_lsn()::text") {
-                    Ok(v) => v,
-                    Err(e) => {
-                        pgrx::warning!(
-                            "pg_trickle scheduler: failed to fetch pg_current_wal_lsn (worker/fallback): {}",
-                            e
-                        );
-                        None
-                    }
-                };
-            }
-
-            // Cap with current write LSN: don't advance past what's available now.
-            let write_lsn_str = match Spi::get_one::<String>("SELECT pg_current_wal_lsn()::text") {
-                Ok(Some(v)) => v,
-                Ok(None) => "0/0".to_string(),
-                Err(e) => {
-                    pgrx::warning!(
-                        "pg_trickle scheduler: failed to fetch pg_current_wal_lsn (worker/cap): {}",
-                        e
-                    );
-                    "0/0".to_string()
-                }
-            };
-            let write_u64 = version::lsn_to_u64(&write_lsn_str);
-            let effective_u64 = safe_lsn_u64.min(write_u64);
-            Some(version::u64_to_lsn(effective_u64))
-        }
-    }
-}
-
-/// Rate-limited WARNING for when frontier holdback has been active longer
-/// than `pg_trickle.frontier_holdback_warn_seconds`.
-///
-/// Emits at most one WARNING per minute.
-fn emit_holdback_warning_if_needed(oldest_txn_age_secs: u64) {
-    let warn_secs = config::pg_trickle_frontier_holdback_warn_seconds();
-    if warn_secs <= 0 {
-        return;
-    }
-    if oldest_txn_age_secs < warn_secs as u64 {
-        return;
-    }
-
-    // Rate-limit: emit at most once per minute.
-    let now_secs = std::time::SystemTime::now()
-        .duration_since(std::time::UNIX_EPOCH)
-        .unwrap_or_default()
-        .as_secs();
-    let last_warn = LAST_HOLDBACK_WARN_SECS.load(std::sync::atomic::Ordering::Relaxed);
-    if now_secs.saturating_sub(last_warn) < 60 {
-        return;
-    }
-    LAST_HOLDBACK_WARN_SECS.store(now_secs, std::sync::atomic::Ordering::Relaxed);
-
-    pgrx::warning!(
-        "pg_trickle: frontier holdback active — the oldest in-progress transaction is {}s old \
-         (threshold: {}s). Stream tables may lag behind. \
-         Check pg_stat_activity for long-running sessions. \
-         To suppress: SET pg_trickle.frontier_holdback_warn_seconds = 0.",
-        oldest_txn_age_secs,
-        warn_secs,
-    );
 }
 
 /// Execute a singleton unit: refresh a single stream table using the existing inline path.
@@ -1666,1948 +857,6 @@ fn execute_worker_fused_chain(job: &SchedulerJob) -> RefreshOutcome {
         job.job_id,
     );
     RefreshOutcome::Success
-}
-
-// ── Parallel Dispatch State (Phase 4) ─────────────────────────────────────
-
-/// DAG-2: Minimum adaptive poll interval (ms) when workers are in-flight.
-const ADAPTIVE_POLL_MIN_MS: u64 = 20;
-/// DAG-2: Maximum adaptive poll interval (ms) — the old fixed cap.
-const ADAPTIVE_POLL_MAX_MS: u64 = 200;
-
-/// DAG-2: Compute the next adaptive poll interval.
-///
-/// Uses exponential backoff: starts at `ADAPTIVE_POLL_MIN_MS` after a worker
-/// completion, doubles each tick with no completion, and caps at
-/// `ADAPTIVE_POLL_MAX_MS`. When no workers are in-flight, returns the full
-/// `base_interval_ms` (the scheduler GUC value).
-///
-/// This is a pure function for unit-testability.
-fn compute_adaptive_poll_ms(
-    current_poll_ms: u64,
-    had_completion: bool,
-    has_inflight: bool,
-    base_interval_ms: u64,
-) -> u64 {
-    if !has_inflight {
-        return base_interval_ms;
-    }
-    if had_completion {
-        return ADAPTIVE_POLL_MIN_MS;
-    }
-    // Exponential backoff: double the current interval, capped.
-    let next = current_poll_ms.saturating_mul(2);
-    std::cmp::min(next, ADAPTIVE_POLL_MAX_MS)
-}
-
-/// Per-unit coordinator state for the parallel dispatch loop.
-#[derive(Debug, Default)]
-struct UnitDispatchState {
-    /// Number of upstream units that must succeed before this unit is ready.
-    remaining_upstreams: usize,
-    /// In-flight job ID, if a worker is currently executing this unit.
-    inflight_job_id: Option<i64>,
-    /// Whether this unit completed successfully in the current cycle.
-    succeeded: bool,
-}
-
-/// Coordinator-side state for the parallel dispatch loop.
-///
-/// Lives outside the SPI transaction and persists across scheduler ticks.
-/// Rebuilt when the DAG version changes.
-struct ParallelDispatchState {
-    /// Per-unit tracking, keyed by ExecutionUnitId.
-    unit_states: HashMap<ExecutionUnitId, UnitDispatchState>,
-    /// Number of refresh workers in-flight for this database coordinator.
-    per_db_inflight: u32,
-    /// DAG version this dispatch state was built from.
-    dag_version: u64,
-    /// The execution unit DAG (rebuilt on DAG version change).
-    eu_dag: Option<ExecutionUnitDag>,
-    /// DAG-2: Current adaptive poll interval (ms). Doubles each tick without
-    /// worker completions; resets to `ADAPTIVE_POLL_MIN_MS` on completion.
-    adaptive_poll_ms: u64,
-    /// DAG-2: Number of worker completions observed in the last dispatch tick.
-    completions_this_tick: u32,
-}
-
-impl ParallelDispatchState {
-    fn new() -> Self {
-        Self {
-            unit_states: HashMap::new(),
-            per_db_inflight: 0,
-            dag_version: 0,
-            eu_dag: None,
-            adaptive_poll_ms: ADAPTIVE_POLL_MIN_MS,
-            completions_this_tick: 0,
-        }
-    }
-
-    /// Whether there are in-flight jobs (for shorter poll interval).
-    fn has_inflight(&self) -> bool {
-        self.per_db_inflight > 0
-    }
-
-    /// Rebuild dispatch state from a fresh EU DAG.
-    fn rebuild(&mut self, eu_dag: ExecutionUnitDag, dag_version: u64) {
-        // Count how many old units had in-flight jobs that will be orphaned
-        // by clearing unit_states.  Subtract these from per_db_inflight
-        // because the orphaned jobs will be reaped by the dead-worker scan
-        // (Step 0) or will complete and be found via catalog poll. Without
-        // this correction, per_db_inflight permanently drifts upward on
-        // each rebuild, eventually blocking all new dispatches.
-        let orphaned_inflight = self
-            .unit_states
-            .values()
-            .filter(|us| us.inflight_job_id.is_some())
-            .count() as u32;
-        self.per_db_inflight = self.per_db_inflight.saturating_sub(orphaned_inflight);
-
-        self.unit_states.clear();
-        self.dag_version = dag_version;
-
-        for unit in eu_dag.units() {
-            let upstream_count = eu_dag.get_upstream_units(unit.id).len();
-            self.unit_states.insert(
-                unit.id,
-                UnitDispatchState {
-                    remaining_upstreams: upstream_count,
-                    inflight_job_id: None,
-                    succeeded: false,
-                },
-            );
-        }
-        self.eu_dag = Some(eu_dag);
-    }
-}
-
-/// Check if an execution unit is due for refresh.
-///
-/// A unit is due if any member ST is due (schedule or upstream changes).
-fn is_unit_due(unit: &ExecutionUnit, dag: &StDag) -> bool {
-    unit.member_pgt_ids.iter().any(|&pgt_id| {
-        load_st_by_id(pgt_id)
-            .map(|st| {
-                (st.status == StStatus::Active || st.status == StStatus::Initializing)
-                    && (check_schedule(&st, dag) || st.needs_reinit)
-            })
-            .unwrap_or(false)
-    })
-}
-
-// ── C3-1: Per-database worker quota helpers ───────────────────────────────
-
-/// C3-1: Sort an execution unit ready queue by dispatch priority.
-///
-/// Priority ordering (lowest number = dispatched first):
-/// 1. `ImmediateClosure` — transactional consistency, must not be delayed.
-/// 2. `AtomicGroup` / `RepeatableReadGroup` — consistency group members.
-/// 3. `Singleton` — normal scheduled stream tables.
-/// 4. `CyclicScc` — fixpoint groups, lowest urgency.
-///
-/// Topological order is preserved within each priority tier because the
-/// input queue was built by iterating `topo_order`.
-///
-/// The secondary sort key is `tier_priority`: tier values should be
-/// `0 = Hot` (highest urgency), `1 = Warm`, `2 = Cold`.  Pass an empty map
-/// when tier information is unavailable (all units treated as Hot).
-fn sort_ready_queue_by_priority(
-    queue: VecDeque<ExecutionUnitId>,
-    eu_dag: &ExecutionUnitDag,
-    tier_priorities: &HashMap<ExecutionUnitId, u8>,
-) -> VecDeque<ExecutionUnitId> {
-    use crate::dag::ExecutionUnitKind;
-    fn unit_priority(kind: ExecutionUnitKind) -> u8 {
-        match kind {
-            ExecutionUnitKind::ImmediateClosure => 0,
-            ExecutionUnitKind::AtomicGroup | ExecutionUnitKind::RepeatableReadGroup => 1,
-            ExecutionUnitKind::FusedChain => 1, // same priority as atomic groups
-            ExecutionUnitKind::Singleton => 2,
-            ExecutionUnitKind::CyclicScc => 3,
-        }
-    }
-    // C3-1: Tier priority: Hot(0) > Warm(1) > Cold(2).  ImmediateClosure EUs
-    // are already priority 0 by kind and do not need a tier secondary key.
-    let default_tier: u8 = 0; // default to Hot when not in map
-    let mut items: Vec<ExecutionUnitId> = queue.into_iter().collect();
-    // stable sort preserves topological order within each priority tier.
-    items.sort_by_key(|&uid| {
-        let (kind_prio, tier_prio) = eu_dag
-            .units()
-            .find(|u| u.id == uid)
-            .map(|u| {
-                let kp = unit_priority(u.kind);
-                // ImmediateClosure ignores tier — it always wins.
-                let tp = if kp == 0 {
-                    0u8
-                } else {
-                    *tier_priorities.get(&uid).unwrap_or(&default_tier)
-                };
-                (kp, tp)
-            })
-            .unwrap_or((u8::MAX, default_tier));
-        (kind_prio, tier_prio)
-    });
-    items.into_iter().collect()
-}
-
-/// C3-1: Compute the "best" (most-urgent) tier priority for an execution unit.
-///
-/// Iterates the unit's member STs, loads each, and returns the minimum tier
-/// priority value (Hot=0, Warm=1, Cold=2).  Returns `0` (Hot) for any unit
-/// whose members cannot be loaded so they are never accidentally deprioritised.
-///
-/// Pure logic in the sense that it only reads catalog data that has already
-/// been loaded; the `load_st_by_id` calls are cheap (SPI reads, already cached
-/// by the catalog layer during this same scheduler tick).
-fn compute_unit_tier_priority(member_pgt_ids: &[i64]) -> u8 {
-    let mut best: u8 = 2; // Cold — start at lowest urgency
-    for &pgt_id in member_pgt_ids {
-        let tier_prio = match load_st_by_id(pgt_id) {
-            Some(st) => match RefreshTier::from_sql_str(&st.refresh_tier) {
-                RefreshTier::Hot => 0,
-                RefreshTier::Warm => 1,
-                RefreshTier::Cold | RefreshTier::Frozen => 2,
-            },
-            None => 0, // default to Hot urgency when ST cannot be loaded
-        };
-        if tier_prio < best {
-            best = tier_prio;
-        }
-        if best == 0 {
-            break; // Hot is the highest urgency; short-circuit
-        }
-    }
-    best
-}
-
-/// Run one tick of the parallel dispatch loop.
-///
-/// Called from the main scheduler loop when `parallel_refresh_mode == On`.
-/// Polls completed jobs, updates readiness, and enqueues new jobs.
-/// Workers to spawn are appended to `pending_spawns`.
-fn parallel_dispatch_tick(
-    state: &mut ParallelDispatchState,
-    dag: &StDag,
-    now_ms: u64,
-    retry_states: &mut HashMap<i64, RetryState>,
-    retry_policy: &RetryPolicy,
-    db_name: &str,
-    pending_spawns: &mut Vec<(String, i64)>,
-) {
-    let eu_dag = match &state.eu_dag {
-        Some(d) => d,
-        None => return,
-    };
-
-    let max_cluster = config::pg_trickle_max_dynamic_refresh_workers().max(1) as u32;
-    // PAR-2: If max_parallel_workers is set (> 0), use it as an additional cap.
-    let par_workers = config::pg_trickle_max_parallel_workers();
-    let effective_max_cluster = if par_workers > 0 {
-        max_cluster.min(par_workers as u32)
-    } else {
-        max_cluster
-    };
-    // C3-1: Per-database quota with burst capacity.
-    let max_per_db = compute_per_db_quota(
-        config::pg_trickle_per_database_worker_quota(),
-        config::pg_trickle_max_concurrent_refreshes(),
-        effective_max_cluster,
-        shmem::active_worker_count(),
-    );
-    let dag_version_i64 = state.dag_version as i64;
-    // SAFETY: MyProcPid is always valid inside a background worker.
-    let scheduler_pid: i32 = unsafe { pg_sys::MyProcPid };
-
-    // ── Step 0: Reap orphaned RUNNING jobs whose worker died ─────────────
-    // A background worker can die (OOM, crash, etc.) while its job is still
-    // RUNNING in the catalog. The in-memory inflight tracking may have been
-    // lost during a DAG rebuild, so we do a DB-level scan: cancel any
-    // RUNNING job whose worker_pid is no longer in pg_stat_activity.
-    let reaped = reap_dead_worker_jobs();
-    if reaped > 0 {
-        log!(
-            "pg_trickle: parallel dispatch — reaped {} orphaned job(s) with dead workers",
-            reaped,
-        );
-
-        // Reconcile shmem worker token counter.  Crashed workers never call
-        // `release_worker_token()`, so the atomic counter drifts upward.
-        // Re-derive from pg_stat_activity to fix it.
-        let live_workers: u32 = Spi::get_one::<i64>(
-            "SELECT COUNT(*)::bigint FROM pg_stat_activity \
-             WHERE backend_type = 'pg_trickle refresh worker'",
-        )
-        .unwrap_or(Some(0))
-        .unwrap_or(0)
-        .max(0) as u32;
-
-        let shmem_count = shmem::active_worker_count();
-        if shmem_count != live_workers {
-            log!(
-                "pg_trickle: parallel dispatch — correcting worker count: shmem={} → live={}",
-                shmem_count,
-                live_workers,
-            );
-            shmem::set_active_worker_count(live_workers);
-            shmem::bump_reconcile_epoch();
-        }
-    }
-
-    // ── Step 1: Poll completed jobs and process outcomes ──────────────────
-    let inflight_ids: Vec<(ExecutionUnitId, i64)> = state
-        .unit_states
-        .iter()
-        .filter_map(|(&uid, us)| us.inflight_job_id.map(|jid| (uid, jid)))
-        .collect();
-
-    for (unit_id, job_id) in inflight_ids {
-        let job = match SchedulerJob::get_by_id(job_id) {
-            Ok(Some(j)) => j,
-            Ok(None) => {
-                log!(
-                    "pg_trickle: parallel dispatch — job {} vanished, treating as failure",
-                    job_id,
-                );
-                if let Some(us) = state.unit_states.get_mut(&unit_id) {
-                    us.inflight_job_id = None;
-                }
-                state.per_db_inflight = state.per_db_inflight.saturating_sub(1);
-                continue;
-            }
-            Err(e) => {
-                log!(
-                    "pg_trickle: parallel dispatch — error polling job {}: {}",
-                    job_id,
-                    e,
-                );
-                continue; // Try again next tick
-            }
-        };
-
-        if !job.status.is_terminal() {
-            continue; // Still running
-        }
-
-        // Job completed — process outcome.
-        if let Some(us) = state.unit_states.get_mut(&unit_id) {
-            us.inflight_job_id = None;
-        }
-        state.per_db_inflight = state.per_db_inflight.saturating_sub(1);
-        // DAG-2: Track completion for adaptive poll reset.
-        state.completions_this_tick += 1;
-
-        let root_pgt_id = eu_dag
-            .units()
-            .find(|u| u.id == unit_id)
-            .map(|u| u.root_pgt_id);
-
-        match job.status {
-            JobStatus::Succeeded => {
-                if let Some(us) = state.unit_states.get_mut(&unit_id) {
-                    us.succeeded = true;
-                }
-                if let Some(rpid) = root_pgt_id {
-                    retry_states.entry(rpid).or_default().reset();
-                }
-
-                // Advance downstream readiness.
-                for ds_id in eu_dag.get_downstream_units(unit_id) {
-                    if let Some(ds_state) = state.unit_states.get_mut(&ds_id) {
-                        ds_state.remaining_upstreams =
-                            ds_state.remaining_upstreams.saturating_sub(1);
-                    }
-                }
-
-                log!(
-                    "pg_trickle: parallel dispatch — unit completed (job {})",
-                    job_id,
-                );
-            }
-            JobStatus::RetryableFailed => {
-                if let Some(rpid) = root_pgt_id {
-                    let retry = retry_states.entry(rpid).or_default();
-                    let will_retry = retry.record_failure(retry_policy, now_ms);
-                    if will_retry {
-                        log!(
-                            "pg_trickle: parallel dispatch — job {} failed (retryable), backoff {}ms",
-                            job_id,
-                            retry_policy.backoff_ms(retry.attempts - 1),
-                        );
-                    } else {
-                        // Retry budget exhausted — unblock downstream units
-                        // so the wave can complete. Without this, downstreams
-                        // remain permanently blocked because the failed unit
-                        // never sets succeeded=true and remaining_upstreams
-                        // is never decremented.
-                        log!(
-                            "pg_trickle: parallel dispatch — job {} retries exhausted, unblocking downstreams",
-                            job_id,
-                        );
-                        if let Some(us) = state.unit_states.get_mut(&unit_id) {
-                            us.succeeded = true;
-                        }
-                        for ds_id in eu_dag.get_downstream_units(unit_id) {
-                            if let Some(ds_state) = state.unit_states.get_mut(&ds_id) {
-                                ds_state.remaining_upstreams =
-                                    ds_state.remaining_upstreams.saturating_sub(1);
-                            }
-                        }
-                    }
-                }
-            }
-            JobStatus::PermanentFailed | JobStatus::Cancelled => {
-                if let Some(rpid) = root_pgt_id {
-                    retry_states.entry(rpid).or_default().reset();
-                }
-                log!(
-                    "pg_trickle: parallel dispatch — job {} failed permanently",
-                    job_id,
-                );
-
-                // ERR-1d: Ensure ERROR status is set on member STs for permanent
-                // failures. The worker may have already set this via the
-                // catch_unwind path, but this is a safety net in case the
-                // worker died before completing the error-state UPDATE.
-                if job.status == JobStatus::PermanentFailed {
-                    // PERF-5: O(1) lookup via unit_by_id.
-                    let unit = eu_dag.unit_by_id(unit_id);
-                    if let Some(unit) = unit {
-                        let error_detail = job
-                            .outcome_detail
-                            .as_deref()
-                            .unwrap_or("Refresh failed permanently");
-                        for &pgt_id in &unit.member_pgt_ids {
-                            // Only set ERROR if not already in ERROR state.
-                            if let Some(st) = load_st_by_id(pgt_id)
-                                && st.status != StStatus::Error
-                            {
-                                let _ = StreamTableMeta::set_error_state(pgt_id, error_detail);
-                            }
-                        }
-                    }
-                }
-
-                // Mark this unit as "succeeded" so the wave can complete.
-                // The refresh itself already recorded the failure in
-                // pgt_refresh_history, but we must unblock downstream
-                // units — otherwise a single permanent failure silently
-                // stalls every transitive downstream for the remainder of
-                // the scheduler's lifetime.
-                if let Some(us) = state.unit_states.get_mut(&unit_id) {
-                    us.succeeded = true;
-                }
-                for ds_id in eu_dag.get_downstream_units(unit_id) {
-                    if let Some(ds_state) = state.unit_states.get_mut(&ds_id) {
-                        ds_state.remaining_upstreams =
-                            ds_state.remaining_upstreams.saturating_sub(1);
-                    }
-                }
-            }
-            _ => {}
-        }
-    }
-
-    // ── Step 2: Build ready queue ────────────────────────────────────────
-    let mut ready_queue: VecDeque<ExecutionUnitId> = VecDeque::new();
-
-    // Use topological order for deterministic upstream-first priority.
-    let topo_order = match eu_dag.topological_order() {
-        Ok(order) => order,
-        Err(e) => {
-            log!("pg_trickle: parallel dispatch — topo order failed: {}", e);
-            return;
-        }
-    };
-
-    for &uid in &topo_order {
-        let us = match state.unit_states.get(&uid) {
-            Some(s) => s,
-            None => continue,
-        };
-
-        // Skip completed, in-flight, or dependency-blocked units.
-        if us.succeeded || us.inflight_job_id.is_some() || us.remaining_upstreams > 0 {
-            continue;
-        }
-
-        // PERF-5: O(1) lookup via unit_by_id.
-        let unit = match eu_dag.unit_by_id(uid) {
-            Some(u) => u,
-            None => continue,
-        };
-
-        // Check retry backoff.
-        let retry = retry_states.entry(unit.root_pgt_id).or_default();
-        if retry.is_in_backoff(now_ms) {
-            // Emit stale alerts for members in backoff.
-            for &pgt_id in &unit.member_pgt_ids {
-                if let Some(st) = load_st_by_id(pgt_id) {
-                    emit_stale_alert_if_needed(&st);
-                }
-            }
-            continue;
-        }
-
-        // Check if unit is due for refresh.
-        if !is_unit_due(unit, dag) {
-            continue;
-        }
-
-        // Guard against duplicate in-flight jobs.
-        match SchedulerJob::has_inflight_job(&unit.stable_key()) {
-            Ok(true) => continue,
-            Ok(false) => {}
-            Err(e) => {
-                log!(
-                    "pg_trickle: parallel dispatch — inflight check error for {}: {}",
-                    unit.label,
-                    e,
-                );
-                continue;
-            }
-        }
-
-        ready_queue.push_back(uid);
-    }
-
-    // C3-1: Build tier priority map for the ready queue.
-    // For each EU in the ready queue, compute the "best" (most urgent) tier
-    // among its member STs.  This is the secondary sort key used below.
-    // We only compute for units that are actually ready so the map stays small.
-    // PERF-5: O(1) lookup via unit_by_id.
-    let tier_map: HashMap<ExecutionUnitId, u8> = ready_queue
-        .iter()
-        .filter_map(|&uid| {
-            eu_dag
-                .unit_by_id(uid)
-                .map(|u| (uid, compute_unit_tier_priority(&u.member_pgt_ids)))
-        })
-        .collect();
-
-    // C3-1: Priority sort — IMMEDIATE closures first for transactional safety,
-    // then atomic groups, singletons (Hot > Warm > Cold), cyclic SCCs.
-    // Topological order within each priority tier is preserved.
-    let ready_queue = sort_ready_queue_by_priority(ready_queue, eu_dag, &tier_map);
-
-    // ── Step 3: Dispatch ready units within budget ───────────────────────
-    let mut ready_queue = ready_queue;
-    while let Some(uid) = ready_queue.pop_front() {
-        if state.per_db_inflight >= max_per_db {
-            break;
-        }
-
-        if !shmem::try_acquire_worker_token(max_cluster) {
-            log!(
-                "pg_trickle: parallel dispatch — worker budget exhausted ({}/{})",
-                shmem::active_worker_count(),
-                max_cluster,
-            );
-            break;
-        }
-
-        // PERF-5: O(1) lookup via unit_by_id.
-        let unit = match eu_dag.unit_by_id(uid) {
-            Some(u) => u,
-            None => {
-                shmem::release_worker_token();
-                continue;
-            }
-        };
-
-        let job_id = match SchedulerJob::enqueue(
-            dag_version_i64,
-            &unit.stable_key(),
-            unit.kind.as_str(),
-            &unit.member_pgt_ids,
-            unit.root_pgt_id,
-            scheduler_pid,
-            1,
-        ) {
-            Ok(id) => id,
-            Err(e) => {
-                shmem::release_worker_token();
-                log!(
-                    "pg_trickle: parallel dispatch — failed to enqueue job for {}: {}",
-                    unit.label,
-                    e,
-                );
-                continue;
-            }
-        };
-
-        if let Some(us) = state.unit_states.get_mut(&uid) {
-            us.inflight_job_id = Some(job_id);
-        }
-        state.per_db_inflight += 1;
-        pending_spawns.push((db_name.to_string(), job_id));
-
-        log!(
-            "pg_trickle: parallel dispatch — enqueued {} (job_id={}, kind={})",
-            unit.label,
-            job_id,
-            unit.kind,
-        );
-    }
-
-    // ── Step 4: Reset cycle state when the wave is fully complete ────────
-    // A dispatch wave completes when all in-flight workers have finished AND
-    // nothing new was dispatched in Step 3 (per_db_inflight is still 0 after
-    // the dispatch loop).  At that point `succeeded = true` on every unit
-    // that ran this wave, and we can clear those flags so units become
-    // eligible for the next scheduled wave.
-    //
-    // Placing the reset HERE (after dispatch) rather than before Step 2 is
-    // critical for cascade correctness.  With a two-unit cascade A → B:
-    //   • A completes in Step 1 → per_db_inflight drops to 0, B.remaining=0.
-    //   • Step 2 sees B ready; Step 3 dispatches B → per_db_inflight becomes 1.
-    //   • Reset check: per_db_inflight == 1 → no reset.  B runs correctly.
-    // If the reset were placed before Step 2, per_db_inflight would be 0 at
-    // that point (B not yet dispatched), causing B.remaining_upstreams to be
-    // restored to 1 — blocking B forever in a cascade.
-    if state.per_db_inflight == 0 {
-        let any_done = state.unit_states.values().any(|us| us.succeeded);
-        if any_done {
-            for (&uid, us) in state.unit_states.iter_mut() {
-                us.succeeded = false;
-                us.remaining_upstreams = eu_dag.get_upstream_units(uid).len();
-            }
-            log!(
-                "pg_trickle: parallel dispatch — wave complete, reset {} unit(s)",
-                state.unit_states.len()
-            );
-        }
-    }
-}
-
-/// Reap RUNNING jobs whose background worker process has died.
-///
-/// Called at the start of each parallel dispatch tick to detect orphaned
-/// RUNNING jobs whose worker_pid is no longer in pg_stat_activity.
-/// This handles the case where a worker crashes (OOM, segfault, etc.)
-/// while its job status remains RUNNING in the catalog — which would
-/// otherwise permanently block scheduling for that execution unit.
-///
-/// Returns the number of jobs reaped.
-fn reap_dead_worker_jobs() -> i64 {
-    Spi::get_one::<i64>(
-        "WITH reaped AS ( \
-             UPDATE pgtrickle.pgt_scheduler_jobs \
-             SET status = 'RETRYABLE_FAILED', \
-                 finished_at = now(), \
-                 outcome_detail = 'Worker process died (reaped by scheduler)', \
-                 retryable = true \
-             WHERE status = 'RUNNING' \
-               AND worker_pid IS NOT NULL \
-               AND NOT EXISTS ( \
-                   SELECT 1 FROM pg_stat_activity \
-                   WHERE pid = pgt_scheduler_jobs.worker_pid \
-               ) \
-             RETURNING job_id \
-         ) SELECT count(*)::bigint FROM reaped",
-    )
-    .unwrap_or(Some(0))
-    .unwrap_or(0)
-}
-
-/// Reconcile orphaned jobs and worker tokens at scheduler startup.
-///
-/// Called once during per-database scheduler initialization:
-/// 1. Cancels orphaned QUEUED/RUNNING jobs whose PID is no longer alive.
-/// 2. Computes the true active worker count from `pg_stat_activity`.
-/// 3. Corrects the shared-memory token counter if it diverged.
-pub fn reconcile_parallel_state() {
-    // Step 1: Cancel orphaned jobs
-    match SchedulerJob::cancel_orphaned_jobs() {
-        Ok(count) if count > 0 => {
-            log!(
-                "pg_trickle: parallel reconciliation — cancelled {} orphaned job(s)",
-                count
-            );
-        }
-        Ok(_) => {}
-        Err(e) => {
-            log!(
-                "pg_trickle: parallel reconciliation — orphan cleanup error: {}",
-                e
-            );
-        }
-    }
-
-    // Step 2: Count live refresh workers from pg_stat_activity
-    let live_workers: u32 = Spi::get_one::<i64>(
-        "SELECT COUNT(*)::bigint FROM pg_stat_activity \
-         WHERE backend_type = 'pg_trickle refresh worker'",
-    )
-    .unwrap_or(Some(0))
-    .unwrap_or(0)
-    .max(0) as u32;
-
-    // Step 3: Correct shared-memory counter if needed
-    let shmem_count = shmem::active_worker_count();
-    if shmem_count != live_workers {
-        log!(
-            "pg_trickle: parallel reconciliation — correcting worker count: shmem={} → live={}",
-            shmem_count,
-            live_workers,
-        );
-        shmem::set_active_worker_count(live_workers);
-        shmem::bump_reconcile_epoch();
-    }
-
-    // Step 4: Prune old completed jobs (keep last 1 hour)
-    match SchedulerJob::prune_completed(3600) {
-        Ok(count) if count > 0 => {
-            log!(
-                "pg_trickle: parallel reconciliation — pruned {} old job(s)",
-                count
-            );
-        }
-        Ok(_) => {}
-        Err(e) => {
-            log!("pg_trickle: parallel reconciliation — prune error: {}", e);
-        }
-    }
-}
-
-/// Main entry point for the scheduler background worker.
-///
-/// # Safety
-/// This function is called directly by PostgreSQL as a background worker
-/// entry point. It must follow the C-unwind calling convention.
-#[pg_guard]
-#[unsafe(no_mangle)]
-pub extern "C-unwind" fn pg_trickle_scheduler_main(_arg: pg_sys::Datum) {
-    BackgroundWorker::attach_signal_handlers(SignalWakeFlags::SIGHUP | SignalWakeFlags::SIGTERM);
-
-    // Determine which database to connect to.
-    // Dynamic workers spawned by the launcher have the DB name in bgw_extra.
-    let extra = BackgroundWorker::get_extra();
-    let db_name = if extra.is_empty() {
-        "postgres".to_string()
-    } else {
-        extra.to_string()
-    };
-    BackgroundWorker::connect_worker_to_spi(Some(db_name.as_str()), None);
-
-    // Exit cleanly if pg_trickle is not installed in this database.
-    // The launcher will re-probe after its skip TTL (5 min) expires.
-    let is_installed = BackgroundWorker::transaction(AssertUnwindSafe(|| -> bool {
-        Spi::get_one::<bool>(
-            "SELECT EXISTS(SELECT 1 FROM pg_extension WHERE extname = 'pg_trickle')",
-        )
-        .unwrap_or(Some(false))
-        .unwrap_or(false)
-    }));
-    if !is_installed {
-        log!(
-            "pg_trickle scheduler: pg_trickle not installed in database '{}', exiting",
-            db_name
-        );
-        return;
-    }
-
-    // UG1: Version mismatch check — warn if the compiled .so version differs
-    // from the SQL-installed extension version (stale install).
-    BackgroundWorker::transaction(AssertUnwindSafe(|| {
-        check_extension_version_match();
-    }));
-
-    // F16 (G8.2): Detect read replicas — the scheduler cannot write on a
-    // standby. Skip all work and sleep until promotion.
-    let is_replica = BackgroundWorker::transaction(AssertUnwindSafe(|| -> bool {
-        Spi::get_one::<bool>("SELECT pg_is_in_recovery()")
-            .unwrap_or(Some(false))
-            .unwrap_or(false)
-    }));
-    if is_replica {
-        log!(
-            "pg_trickle scheduler: running on a read replica (pg_is_in_recovery() = true). \
-             Scheduler will sleep until promotion."
-        );
-        // Sleep in a loop — if the server is promoted, pg_is_in_recovery()
-        // changes to false and we can start working.
-        loop {
-            let should_continue =
-                BackgroundWorker::wait_latch(Some(std::time::Duration::from_secs(30)));
-            if !should_continue {
-                log!("pg_trickle scheduler shutting down (replica)");
-                return;
-            }
-            let still_replica = BackgroundWorker::transaction(AssertUnwindSafe(|| -> bool {
-                Spi::get_one::<bool>("SELECT pg_is_in_recovery()")
-                    .unwrap_or(Some(false))
-                    .unwrap_or(false)
-            }));
-            if !still_replica {
-                log!("pg_trickle scheduler: replica promoted — starting normal operation");
-                break;
-            }
-        }
-    }
-
-    log!(
-        "pg_trickle scheduler started (interval={}ms, event_driven_wake={})",
-        config::pg_trickle_scheduler_interval_ms(),
-        config::pg_trickle_event_driven_wake(),
-    );
-
-    // Mark scheduler as running in shared memory so cluster_worker_summary()
-    // and other shmem-based health consumers reflect the correct state.
-    // SAFETY: MyProcPid is always valid inside a background worker.
-    let my_pid = unsafe { pg_sys::MyProcPid };
-    let now_ts = std::time::SystemTime::now()
-        .duration_since(std::time::UNIX_EPOCH)
-        .unwrap_or_default()
-        .as_secs() as i64;
-    crate::shmem::set_scheduler_meta(my_pid, true, now_ts);
-
-    let mut dag_version: u64 = 0;
-    let mut dag: Option<StDag> = None;
-    // Follow-up flag: after an incremental rebuild, schedule a full rebuild
-    // on the next tick.  ALTER operations change dependency edges but not
-    // the node set, so the stale-snapshot guard (which only checks for
-    // missing nodes) cannot detect stale edges.  A follow-up full rebuild
-    // with a fresh snapshot captures the committed edge changes.
-    let mut pending_full_rebuild = false;
-
-    // #536: Previous tick's safe frontier watermark, used by the holdback
-    // algorithm to determine whether any long-running transaction spans a
-    // tick boundary.  Seeded from shared memory on scheduler startup so
-    // that the first post-restart tick preserves the last known-safe LSN
-    // (avoids a one-tick window where the frontier could advance past an
-    // in-flight transaction that was already open before the restart).
-    let mut prev_tick_watermark: Option<String> = {
-        let last_safe = crate::shmem::last_tick_safe_lsn_u64();
-        if last_safe != 0 {
-            Some(crate::version::u64_to_lsn(last_safe))
-        } else {
-            None
-        }
-    };
-
-    // Per-ST retry state (in-memory only, reset on scheduler restart)
-    let mut retry_states: HashMap<i64, RetryState> = HashMap::new();
-    let retry_policy = RetryPolicy::default();
-
-    // Phase 4: Parallel dispatch state (persisted across ticks)
-    let mut parallel_state = ParallelDispatchState::new();
-
-    // Per-ST drift reset counters (differential cycles since last reinit)
-    let mut drift_counters: HashMap<i64, i32> = HashMap::new();
-
-    // P3-5: Per-ST auto-backoff factors (multiplicative schedule stretch).
-    // When auto_backoff is enabled and a ST is falling behind, the factor
-    // doubles each consecutive cycle; resets to 1.0 on the first on-time cycle.
-    let mut backoff_factors: HashMap<i64, f64> = HashMap::new();
-
-    // PH-E2: Per-ST consecutive spill counters.
-    // When a differential refresh writes temp blocks exceeding the threshold,
-    // the counter increments.  After spill_consecutive_limit consecutive spills,
-    // the scheduler forces a FULL refresh.  Resets on any non-spilling refresh.
-    let mut spill_counters: HashMap<i64, i32> = HashMap::new();
-
-    // Timestamp for periodic CDC trigger health check (~every 60s).
-    let mut last_trigger_health_ms: u64 = 0;
-
-    // WM-7: Timestamp for periodic stuck-watermark alerting (~every 60s).
-    let mut last_watermark_stuck_check_ms: u64 = 0;
-    // WM-7: Track source OIDs that have already been reported as stuck to
-    // avoid spamming the NOTIFY channel every check cycle.
-    let mut reported_stuck_sources: HashSet<u32> = HashSet::new();
-
-    // SCAL-1 (v0.31.0): Per-source consecutive-cycle counter for back-pressure
-    // alerting. Maps source_relid (u32) to the number of consecutive scheduler
-    // ticks where that source's change buffer has exceeded the alert threshold.
-    let mut backpressure_cycles: HashMap<u32, i32> = HashMap::new();
-
-    // DB-5: Timestamp for daily history retention cleanup.
-    let mut last_history_cleanup_ms: u64 = 0;
-    const HISTORY_CLEANUP_INTERVAL_MS: u64 = 24 * 60 * 60 * 1000; // 24 hours
-
-    // DF-G2: Dog-feeding auto-apply — timestamp-gated, rate-limited.
-    let mut last_auto_apply_ms: u64 = 0;
-    const AUTO_APPLY_INTERVAL_MS: u64 = 10 * 60 * 1000; // 10 minutes
-
-    // Phase 10: Crash recovery — mark any interrupted RUNNING records
-    BackgroundWorker::transaction(AssertUnwindSafe(|| {
-        recover_from_crash();
-    }));
-
-    // Phase 2: Reconcile parallel refresh state (orphaned jobs, worker tokens)
-    BackgroundWorker::transaction(AssertUnwindSafe(|| {
-        reconcile_parallel_state();
-    }));
-
-    // EC-20: Post-restart CDC TRANSITIONING health check.
-    // If the scheduler was restarted during an active TRIGGER→WAL transition,
-    // verify the transition is still valid. Invalid transitions (stale slot,
-    // missing publication) are rolled back to TRIGGER mode.
-    BackgroundWorker::transaction(AssertUnwindSafe(|| {
-        check_cdc_transition_health();
-    }));
-
-    // WAKE-1: Event-driven wake via LISTEN/NOTIFY.
-    //
-    // PostgreSQL's LISTEN command is restricted to regular backends
-    // (MyBackendType == B_BACKEND). Background workers — which the scheduler
-    // always is — are rejected with "cannot execute LISTEN within a background
-    // process" (async.c:Async_Listen()). Attempting LISTEN causes an elog(ERROR)
-    // that escapes pgrx's catch_unwind and exits the process with exit code 1.
-    //
-    // Until a background-worker-compatible notification mechanism is implemented
-    // (e.g., direct latch signalling via shared memory), event_driven_wake is
-    // always forced to false here regardless of the GUC value. CDC triggers still
-    // emit pg_notify('pgtrickle_wake') for future use once this is re-enabled.
-    let event_driven = false;
-    if config::pg_trickle_event_driven_wake() {
-        // GUC is on but feature is not yet functional in BGWs — warn once at startup.
-        warning!(
-            "pg_trickle scheduler: event_driven_wake=on is not supported in background \
-             workers (PostgreSQL LISTEN is restricted to B_BACKEND processes). \
-             Operating in polling-only mode. Set pg_trickle.event_driven_wake=off \
-             to suppress this warning."
-        );
-    }
-
-    // WAKE-1: Statistics for event-driven vs poll-based wakes.
-    let mut wake_stats_event: u64 = 0;
-    let mut wake_stats_poll: u64 = 0;
-    let mut wake_stats_last_log_ms: u64 = current_epoch_ms();
-
-    // OPS-6: Workload-aware poll — overlap count from df_scheduling_interference.
-    // Refreshed once per auto-apply cycle (10 min).
-    let mut interference_overlap_count: i64 = 0;
-
-    // OP-2: Start the Prometheus metrics HTTP server if metrics_port is non-zero.
-    let metrics_server = {
-        let port = config::pg_trickle_metrics_port();
-        if port > 0 {
-            crate::metrics_server::MetricsServer::start(port as u16)
-        } else {
-            None
-        }
-    };
-
-    loop {
-        // DAG-2: Adaptive poll interval — exponential backoff (20ms → 200ms)
-        // that resets to 20ms on worker completion, making parallel mode
-        // competitive for cheap refreshes.
-        let mut base_interval_ms = config::pg_trickle_scheduler_interval_ms() as u64;
-
-        // OPS-6: Workload-aware poll — if df_scheduling_interference detects
-        // heavy overlap, slightly increase the base interval to reduce
-        // contention. This is a gentle back-off: +10% per overlap pair,
-        // capped at 2× the configured interval.
-        if interference_overlap_count > 0 {
-            let boost = base_interval_ms / 10 * (interference_overlap_count as u64).min(10);
-            base_interval_ms = (base_interval_ms + boost).min(base_interval_ms * 2);
-        }
-
-        let poll_ms = compute_adaptive_poll_ms(
-            parallel_state.adaptive_poll_ms,
-            parallel_state.completions_this_tick > 0,
-            parallel_state.has_inflight(),
-            base_interval_ms,
-        );
-        parallel_state.adaptive_poll_ms = poll_ms;
-        parallel_state.completions_this_tick = 0;
-
-        let wake_start = std::time::Instant::now();
-        let should_continue =
-            BackgroundWorker::wait_latch(Some(std::time::Duration::from_millis(poll_ms)));
-
-        // OP-2: Service one pending Prometheus scrape request per tick (non-blocking).
-        if let Some(ref ms) = metrics_server {
-            // Collect metrics inside a transaction so SPI queries work.
-            let metrics_text = BackgroundWorker::transaction(std::panic::AssertUnwindSafe(|| {
-                crate::monitor::collect_metrics_text()
-            }));
-            ms.serve_one_request(&metrics_text);
-        }
-
-        // WAKE-1: Determine whether this wake was event-driven (notification)
-        // or poll-based (timeout expired). If the latch returned faster than
-        // the poll interval, a notification (or signal) woke us early.
-        let wake_elapsed_ms = wake_start.elapsed().as_millis() as u64;
-        let was_event_wake = event_driven && wake_elapsed_ms < poll_ms.saturating_sub(5);
-
-        if was_event_wake {
-            wake_stats_event += 1;
-            // WAKE-1: Debounce — wait briefly to coalesce rapidly arriving
-            // notifications from bulk DML before starting the tick.
-            let debounce_ms = config::pg_trickle_wake_debounce_ms() as u64;
-            if debounce_ms > 0 {
-                let _ = BackgroundWorker::wait_latch(Some(std::time::Duration::from_millis(
-                    debounce_ms,
-                )));
-            }
-        } else {
-            wake_stats_poll += 1;
-        }
-
-        // WAKE-1: Log wake statistics every 60 seconds.
-        let now_for_stats = current_epoch_ms();
-        if now_for_stats.saturating_sub(wake_stats_last_log_ms) >= 60_000 {
-            if event_driven && (wake_stats_event > 0 || wake_stats_poll > 0) {
-                log!(
-                    "pg_trickle scheduler: wake stats — event={}, poll={} (last 60s)",
-                    wake_stats_event,
-                    wake_stats_poll,
-                );
-            }
-            wake_stats_event = 0;
-            wake_stats_poll = 0;
-            wake_stats_last_log_ms = now_for_stats;
-
-            // Update the last-wake timestamp in shared memory for monitoring.
-            // SAFETY: MyProcPid is always valid inside a background worker.
-            let my_pid = unsafe { pg_sys::MyProcPid };
-            crate::shmem::set_scheduler_meta(my_pid, true, (now_for_stats / 1000) as i64);
-        }
-
-        unsafe {
-            if pg_sys::ConfigReloadPending != 0 {
-                pg_sys::ConfigReloadPending = 0;
-                pg_sys::ProcessConfigFile(pg_sys::GucContext::PGC_SIGHUP);
-                pgrx::info!(
-                    "pg_trickle scheduler: SIGHUP processed, allow_circular is now {}",
-                    config::pg_trickle_allow_circular()
-                );
-            }
-        }
-
-        if !should_continue {
-            // SIGTERM received — shut down gracefully.
-            log!("pg_trickle scheduler shutting down");
-            crate::shmem::set_scheduler_meta(0, false, 0);
-            break;
-        }
-
-        // A35 (v0.36.0): Process pending drain requests on every tick — even
-        // when the scheduler is disabled — so that pgtrickle.drain() never
-        // hangs indefinitely.  When a drain is requested and no refresh workers
-        // are active, this marks the drain as completed (DRAIN_COMPLETED =
-        // DRAIN_REQUESTED) and returns true.  We skip new dispatches for that
-        // tick so callers see a clean quiesced state before drain() returns.
-        if crate::shmem::scheduler_check_and_complete_drain() {
-            continue;
-        }
-
-        if !config::pg_trickle_enabled() {
-            continue;
-        }
-
-        // Check that pg_trickle is still installed in this database.  It can
-        // be dropped at any time via DROP EXTENSION, which removes the schema
-        // and all catalog tables.  Without this guard the very next SPI query
-        // that references pgtrickle.* would crash the worker with exit code 1,
-        // causing the launcher to apply the full 5-minute skip_ttl backoff.
-        // Exiting cleanly here lets the launcher detect the fresh install
-        // using the shorter retry_ttl for databases that previously had a
-        // running scheduler.
-        let still_installed = BackgroundWorker::transaction(AssertUnwindSafe(|| -> bool {
-            Spi::get_one::<bool>(
-                "SELECT EXISTS(SELECT 1 FROM pg_extension WHERE extname = 'pg_trickle')",
-            )
-            .unwrap_or(Some(false))
-            .unwrap_or(false)
-        }));
-        if !still_installed {
-            log!(
-                "pg_trickle scheduler: pg_trickle was dropped from database '{}', exiting",
-                db_name
-            );
-            crate::shmem::set_scheduler_meta(0, false, 0);
-            return;
-        }
-
-        let now_ms = current_epoch_ms();
-
-        // WAL transition processing — three phases with separate transactions
-        // to ensure slot creation happens in a pristine transaction (no prior
-        // SPI reads that could assign an XID via hint-bit WAL writes).
-        //
-        // Phase 1: Check eligibility, poll WAL sources, collect pending slots
-        // Phase 2: Create replication slots (NO SPI — pristine transaction)
-        // Phase 3: Finish transitions (publication + catalog update)
-        let mut pending_slots = Vec::new();
-        let mut pending_aborts = Vec::new();
-        BackgroundWorker::transaction(AssertUnwindSafe(|| {
-            let change_schema = config::pg_trickle_change_buffer_schema();
-            match wal_decoder::advance_wal_transitions_phase1(&change_schema) {
-                Ok(result) => {
-                    pending_slots = result.pending_slots;
-                    pending_aborts = result.pending_aborts;
-                }
-                Err(e) => log!("pg_trickle: WAL transition phase 1 error: {}", e),
-            }
-            // NOTE: check_slot_health_and_alert() is intentionally NOT called
-            // here.  Phase 1's WAL poll for a missing/invalid replication slot
-            // can abort the current SPI session (even when the Rust-level
-            // catch_unwind absorbs the panic), which silently breaks subsequent
-            // SPI calls — including the EC-34 slot-existence check inside
-            // check_slot_health_and_alert().  Running the health check in its
-            // own transaction (below) guarantees a pristine SPI connection and
-            // ensures reliable fallback-to-TRIGGER detection.
-        }));
-
-        // Slot / buffer health check in its own pristine transaction.
-        // Must be separate from Phase 1 so that WAL poll errors (which can
-        // corrupt the Phase 1 SPI session) do not prevent the EC-34 missing-
-        // slot detection from running.  This is the primary path for automatic
-        // fallback from WAL mode to TRIGGER mode when a replication slot is
-        // externally dropped.
-        BackgroundWorker::transaction(AssertUnwindSafe(|| {
-            monitor::check_slot_health_and_alert();
-        }));
-
-        // SCAL-1 (v0.31.0): Back-pressure detection — check change buffer sizes
-        // and emit change_buffer_backpressure alerts when buffers are persistently
-        // large across multiple consecutive refresh cycles.
-        {
-            let limit = config::pg_trickle_backpressure_consecutive_limit();
-            if limit > 0 {
-                BackgroundWorker::transaction(AssertUnwindSafe(|| {
-                    let threshold = config::pg_trickle_buffer_alert_threshold();
-                    let over = monitor::check_change_buffer_sizes();
-                    // Update consecutive-cycle counters for each source.
-                    let now_over: std::collections::HashSet<u32> =
-                        over.iter().map(|(oid, _)| *oid).collect();
-                    // Increment counters for sources over threshold.
-                    for (oid, pending) in &over {
-                        let cnt = backpressure_cycles
-                            .entry(*oid)
-                            .and_modify(|c| *c += 1)
-                            .or_insert(1);
-                        if *cnt >= limit {
-                            monitor::alert_change_buffer_backpressure(
-                                *oid, *pending, *cnt, threshold,
-                            );
-                            pgrx::log!(
-                                "[pg_trickle] SCAL-1: back-pressure alert for source OID {} \
-                                 — {} rows over threshold {} for {} consecutive cycles",
-                                oid,
-                                pending,
-                                threshold,
-                                cnt,
-                            );
-                        }
-                    }
-                    // Reset counters for sources that came back under threshold.
-                    backpressure_cycles.retain(|oid, _| now_over.contains(oid));
-                }));
-            }
-        }
-
-        // Periodic CDC trigger health check: detect disabled/missing triggers
-        // on source tables.  Runs every ~60s to avoid per-tick overhead.
-        let now_for_trigger_check = current_epoch_ms();
-        if now_for_trigger_check.saturating_sub(last_trigger_health_ms) >= 60_000 {
-            BackgroundWorker::transaction(AssertUnwindSafe(|| {
-                monitor::check_cdc_trigger_health();
-            }));
-            last_trigger_health_ms = now_for_trigger_check;
-
-            // STAB-3 (v0.30.0): Age-based purge of the L2 catalog template cache.
-            // Runs on the same ~60s cadence as the CDC trigger health check to
-            // avoid adding another time-tracking variable.
-            let max_age_hours = config::pg_trickle_template_cache_max_age_hours();
-            if max_age_hours > 0 {
-                BackgroundWorker::transaction(AssertUnwindSafe(|| {
-                    let purged = crate::template_cache::purge_stale_entries(max_age_hours);
-                    if purged > 0 {
-                        pgrx::log!(
-                            "[pg_trickle] STAB-3: purged {} stale template cache entries \
-                             older than {} hours",
-                            purged,
-                            max_age_hours,
-                        );
-                    }
-                }));
-            }
-        }
-
-        // WM-7: Periodic stuck-watermark detection and alerting (~every 60s).
-        let timeout = config::pg_trickle_watermark_holdback_timeout();
-        if timeout > 0 {
-            let now_for_wm_check = current_epoch_ms();
-            if now_for_wm_check.saturating_sub(last_watermark_stuck_check_ms) >= 60_000 {
-                BackgroundWorker::transaction(AssertUnwindSafe(|| {
-                    match crate::catalog::find_stuck_watermarks(timeout) {
-                        Ok(stuck_list) => {
-                            // Collect currently-stuck OIDs for comparison.
-                            let current_stuck: HashSet<u32> =
-                                stuck_list.iter().map(|(_, oid, _)| *oid).collect();
-
-                            // Emit NOTIFY for newly-stuck sources only.
-                            for (group_name, source_oid, age_secs) in &stuck_list {
-                                if !reported_stuck_sources.contains(source_oid) {
-                                    let payload = format!(
-                                        "{{\"event\":\"watermark_stuck\",\"group\":\"{}\",\
-                                          \"source_oid\":{},\"age_secs\":{:.0}}}",
-                                        group_name, source_oid, age_secs
-                                    );
-                                    let _ = Spi::run_with_args(
-                                        "SELECT pg_notify('pgtrickle_alert', $1)",
-                                        &[payload.as_str().into()],
-                                    );
-                                    pgrx::warning!(
-                                        "pg_trickle: watermark stuck in group '{}' — \
-                                         source OID {} not advanced for {:.0}s (timeout {}s)",
-                                        group_name,
-                                        source_oid,
-                                        age_secs,
-                                        timeout,
-                                    );
-                                }
-                            }
-
-                            // Emit recovery NOTIFY for sources that were stuck but
-                            // have since advanced (auto-resume).
-                            for previously_stuck in &reported_stuck_sources {
-                                if !current_stuck.contains(previously_stuck) {
-                                    let payload = format!(
-                                        "{{\"event\":\"watermark_resumed\",\
-                                          \"source_oid\":{}}}",
-                                        previously_stuck
-                                    );
-                                    let _ = Spi::run_with_args(
-                                        "SELECT pg_notify('pgtrickle_alert', $1)",
-                                        &[payload.as_str().into()],
-                                    );
-                                    pgrx::info!(
-                                        "pg_trickle: watermark resumed for source OID {}",
-                                        previously_stuck,
-                                    );
-                                }
-                            }
-
-                            reported_stuck_sources = current_stuck;
-                        }
-                        Err(e) => {
-                            pgrx::warning!("pg_trickle: stuck watermark check failed: {}", e,);
-                        }
-                    }
-                }));
-                last_watermark_stuck_check_ms = now_for_wm_check;
-            }
-        }
-
-        // DB-5: Periodic history retention cleanup (daily).
-        {
-            let now_for_cleanup = current_epoch_ms();
-            if now_for_cleanup.saturating_sub(last_history_cleanup_ms)
-                >= HISTORY_CLEANUP_INTERVAL_MS
-            {
-                let retention_days = config::pg_trickle_history_retention_days();
-                if retention_days > 0 {
-                    // A10 (v0.35.0): Batched DELETEs — delete up to 10,000 rows per
-                    // transaction to limit lock contention on pgt_refresh_history.
-                    // Repeat until fewer than the batch size are deleted.
-                    let batch_size: i64 = 10_000;
-                    let mut total_deleted: i64 = 0;
-                    loop {
-                        let deleted: i64 = BackgroundWorker::transaction(AssertUnwindSafe(|| {
-                            Spi::get_one_with_args::<i64>(
-                                "WITH batch AS (\
-                                        SELECT refresh_id FROM pgtrickle.pgt_refresh_history \
-                                        WHERE start_time < now() - make_interval(days => $1) \
-                                        LIMIT $2 \
-                                    ), deleted AS (\
-                                        DELETE FROM pgtrickle.pgt_refresh_history \
-                                        WHERE refresh_id IN (SELECT refresh_id FROM batch) \
-                                        RETURNING 1\
-                                    ) SELECT count(*) FROM deleted",
-                                &[retention_days.into(), batch_size.into()],
-                            )
-                            .unwrap_or(Some(0))
-                            .unwrap_or(0)
-                        }));
-                        total_deleted += deleted;
-                        if deleted < batch_size {
-                            break;
-                        }
-                    }
-                    if total_deleted > 0 {
-                        log!(
-                            "pg_trickle: history cleanup — deleted {} rows older than {} days (batched)",
-                            total_deleted,
-                            retention_days,
-                        );
-                    }
-                }
-                last_history_cleanup_ms = now_for_cleanup;
-            }
-        }
-
-        // DF-G2: Dog-feeding auto-apply — read df_threshold_advice, apply changes.
-        {
-            let now_for_auto_apply = current_epoch_ms();
-            if now_for_auto_apply.saturating_sub(last_auto_apply_ms) >= AUTO_APPLY_INTERVAL_MS {
-                let auto_apply_mode = config::pg_trickle_self_monitoring_auto_apply();
-                if auto_apply_mode != config::SelfMonitoringAutoApply::Off {
-                    BackgroundWorker::transaction(AssertUnwindSafe(|| {
-                        self_monitoring_auto_apply_tick();
-                    }));
-                }
-                last_auto_apply_ms = now_for_auto_apply;
-
-                // SLA-3: Dynamic tier re-assignment — check and adjust tiers
-                // for stream tables with SLA configured.
-                BackgroundWorker::transaction(AssertUnwindSafe(|| {
-                    sla_tier_adjustment_tick();
-                }));
-
-                // OPS-6: Refresh interference overlap count for workload-aware poll.
-                // Only reads if the DF ST exists (safe even without self-monitoring).
-                BackgroundWorker::transaction(AssertUnwindSafe(|| {
-                    let si_exists: bool = Spi::get_one(
-                        "SELECT EXISTS ( \
-                            SELECT 1 FROM pgtrickle.pgt_stream_tables \
-                            WHERE pgt_schema = 'pgtrickle' \
-                              AND pgt_name = 'df_scheduling_interference' \
-                        )",
-                    )
-                    .unwrap_or(Some(false))
-                    .unwrap_or(false);
-                    if si_exists {
-                        interference_overlap_count = Spi::get_one(
-                            "SELECT coalesce(sum(overlap_count), 0)::bigint \
-                             FROM pgtrickle.df_scheduling_interference",
-                        )
-                        .unwrap_or(Some(0))
-                        .unwrap_or(0);
-                    } else {
-                        interference_overlap_count = 0;
-                    }
-                }));
-            }
-        }
-
-        // Phase 2: Create each pending slot in its own pristine transaction.
-        // No SPI calls — just the C replication API.
-        let mut created_slots = Vec::new();
-        for pending in pending_slots {
-            let slot_name = pending.slot_name.clone();
-            let mut slot_lsn = None;
-            BackgroundWorker::transaction(AssertUnwindSafe(|| {
-                match wal_decoder::create_replication_slot_pristine(&slot_name) {
-                    Ok(lsn) => {
-                        log!("pg_trickle: created replication slot '{}'", slot_name);
-                        slot_lsn = Some(lsn);
-                    }
-                    Err(e) => {
-                        log!(
-                            "pg_trickle: failed to create replication slot '{}': {}",
-                            slot_name,
-                            e
-                        );
-                    }
-                }
-            }));
-            if let Some(lsn) = slot_lsn {
-                created_slots.push((pending, lsn));
-            }
-        }
-
-        // Phase 3: Finish transitions (publications + catalog updates)
-        if !created_slots.is_empty() {
-            BackgroundWorker::transaction(AssertUnwindSafe(|| {
-                if let Err(e) = wal_decoder::advance_wal_transitions_phase3(&created_slots) {
-                    log!("pg_trickle: WAL transition phase 3 error: {}", e);
-                }
-            }));
-        }
-
-        // Phase 4: Abort WAL transitions that need fallback to triggers.
-        // Each abort runs in its own transaction because Phase 1's SPI may
-        // be broken after a caught panic from a missing slot.
-        for abort in pending_aborts {
-            BackgroundWorker::transaction(AssertUnwindSafe(|| {
-                let change_schema = config::pg_trickle_change_buffer_schema();
-                if let Err(e) = wal_decoder::abort_wal_transition(
-                    abort.source_relid,
-                    abort.pgt_id,
-                    &change_schema,
-                ) {
-                    warning!(
-                        "pg_trickle: WAL abort (fallback to triggers) failed for OID {}: {}",
-                        abort.source_relid.to_u32(),
-                        e
-                    );
-                }
-            }));
-        }
-
-        // Collect jobs to spawn (populated inside the transaction, spawned after).
-        let mut pending_spawns: Vec<(String, i64)> = Vec::new();
-
-        // Run the scheduler tick inside a transaction
-        BackgroundWorker::transaction(AssertUnwindSafe(|| {
-            // CSS1 / #536: Capture tick watermark for cross-source snapshot consistency
-            // with frontier holdback to prevent silent data loss from long-running
-            // transactions that span a tick boundary.
-            let (tick_watermark, _current_oldest_xmin, _holdback_age_secs) =
-                compute_coordinator_tick_watermark(prev_tick_watermark.as_deref());
-            // Persist this tick's safe watermark for the next tick's holdback comparison.
-            prev_tick_watermark.clone_from(&tick_watermark);
-
-            // Step A: Check if DAG needs rebuild
-            let current_version = shmem::current_dag_version();
-            if current_version != dag_version || dag.is_none() || pending_full_rebuild {
-                let force_full = pending_full_rebuild;
-                pending_full_rebuild = false;
-
-                // C2-1: Drain the invalidation ring buffer.
-                let invalidated = shmem::drain_invalidations();
-
-                // G-8: Try incremental rebuild when we have specific pgt_ids
-                // and an existing DAG. Falls back to full rebuild on overflow,
-                // no existing DAG, or incremental failure.
-                // Skip incremental if a follow-up full rebuild was requested.
-                let incremental_ok = if force_full {
-                    false
-                } else if let Some(ids) = &invalidated {
-                    if !ids.is_empty() {
-                        if let Some(ref mut existing_dag) = dag {
-                            match existing_dag.rebuild_incremental(
-                                ids,
-                                config::pg_trickle_default_schedule_seconds(),
-                            ) {
-                                Ok(()) => {
-                                    log!(
-                                        "pg_trickle: DAG incrementally updated for pgt_ids {:?} (version={})",
-                                        ids,
-                                        current_version,
-                                    );
-                                    true
-                                }
-                                Err(e) => {
-                                    log!(
-                                        "pg_trickle: incremental DAG rebuild failed ({}), falling back to full rebuild",
-                                        e,
-                                    );
-                                    false
-                                }
-                            }
-                        } else {
-                            false
-                        }
-                    } else {
-                        false
-                    }
-                } else {
-                    // Overflow → full rebuild
-                    log!("pg_trickle: DAG invalidation overflow → full rebuild");
-                    false
-                };
-
-                if !incremental_ok {
-                    match StDag::build_from_catalog(config::pg_trickle_default_schedule_seconds()) {
-                        Ok(new_dag) => {
-                            dag = Some(new_dag);
-                            log!("pg_trickle: DAG rebuilt (version={})", current_version);
-                        }
-                        Err(e) => {
-                            log!("pg_trickle: failed to rebuild DAG: {}", e);
-                            return;
-                        }
-                    }
-                }
-
-                // Race-condition guard: verify all invalidated pgt_ids are visible
-                // in the rebuilt DAG.  `signal_dag_invalidation` is called inside
-                // the creating transaction before it commits, so the scheduler can
-                // read the new version while the transaction is still in-flight.
-                // If the SPI snapshot captured at tick start pre-dates that commit,
-                // the catalog query misses the new ST and the DAG is built without
-                // it.  Detect this by checking whether every invalidated id is
-                // present; if not, trigger a full-rebuild signal and skip updating
-                // dag_version so the next tick rebuilds with a fresh snapshot that
-                // includes all committed STs.
-                let stale_snapshot_detected = if let (Some(ids), Some(d)) = (&invalidated, &dag) {
-                    ids.iter().any(|&id| !d.has_st_node(id))
-                } else {
-                    false
-                };
-
-                if stale_snapshot_detected {
-                    log!(
-                        "pg_trickle: DAG rebuild used stale snapshot — some invalidated pgt_ids \
-                         not yet visible (version={}); triggering full rebuild next tick",
-                        current_version
-                    );
-                    // Signal a full rebuild (overflow) so the next tick uses a fresh
-                    // snapshot.  This is necessary because without the overflow flag,
-                    // an incremental rebuild for a *different* pgt_id arriving between
-                    // ticks could satisfy the version-mismatch check while still
-                    // missing the not-yet-committed ST — permanently excluding it from
-                    // the DAG.  The overflow path forces a full build_from_catalog()
-                    // on the next tick regardless of what else arrives in the ring.
-                    //
-                    // Do NOT update dag_version — the stale tick's version is not
-                    // trusted.
-                    shmem::signal_dag_rebuild();
-                } else {
-                    dag_version = current_version;
-                    // After an incremental rebuild, schedule a follow-up full
-                    // rebuild to catch edge-stale scenarios.  ALTER operations
-                    // update dependency edges but preserve the node; the stale
-                    // guard above only detects missing *nodes*.  A full rebuild
-                    // on the next tick (with a fresh snapshot that sees the
-                    // committed edge updates) closes this window.
-                    if incremental_ok {
-                        pending_full_rebuild = true;
-                    }
-                }
-            }
-
-            let dag_ref = match &dag {
-                Some(d) => d,
-                None => return,
-            };
-
-            // Step B: Validate topological order (detect cycles)
-            // When circular dependencies are allowed (allow_circular=true),
-            // we use the condensation order (SCC-based) instead of rejecting
-            // cycles outright. Cyclic SCCs are handled by fixpoint iteration.
-            let allow_circular = config::pg_trickle_allow_circular();
-            if !allow_circular && let Err(e) = dag_ref.topological_order() {
-                log!("pg_trickle: DAG has cycles: {}", e);
-                return;
-            }
-
-            // Step B2: Build execution unit DAG for parallel-refresh awareness.
-            let parallel_mode = config::pg_trickle_parallel_refresh_mode();
-            match parallel_mode {
-                config::ParallelRefreshMode::Off => {}
-                config::ParallelRefreshMode::DryRun => {
-                    let eu_dag = ExecutionUnitDag::build_from_st_dag(dag_ref, |pgt_id| {
-                        load_st_by_id(pgt_id).map(|st| st.refresh_mode)
-                    });
-                    log!(
-                        "pg_trickle: parallel refresh (dry_run): {}",
-                        eu_dag.summary()
-                    );
-                    log!("{}", eu_dag.dry_run_log());
-                    // Fall through to sequential refresh.
-                }
-                config::ParallelRefreshMode::On => {
-                    // Phase 4: Parallel dispatch replaces sequential refresh.
-                    if parallel_state.dag_version != dag_version || parallel_state.eu_dag.is_none()
-                    {
-                        let eu_dag = ExecutionUnitDag::build_from_st_dag(dag_ref, |pgt_id| {
-                            load_st_by_id(pgt_id).map(|st| st.refresh_mode)
-                        });
-                        log!(
-                            "pg_trickle: parallel dispatch — EU DAG rebuilt: {}",
-                            eu_dag.summary()
-                        );
-                        parallel_state.rebuild(eu_dag, dag_version);
-                    }
-
-                    parallel_dispatch_tick(
-                        &mut parallel_state,
-                        dag_ref,
-                        now_ms,
-                        &mut retry_states,
-                        &retry_policy,
-                        &db_name,
-                        &mut pending_spawns,
-                    );
-
-                    // Prune retry states for STs that no longer exist.
-                    if let Some(ref eu) = parallel_state.eu_dag {
-                        let active_ids: HashSet<i64> = eu
-                            .units()
-                            .flat_map(|u| u.member_pgt_ids.iter().copied())
-                            .collect();
-                        retry_states.retain(|id, _| active_ids.contains(id));
-                    }
-
-                    return; // Skip sequential refresh.
-                }
-            }
-
-            // Step B3: Handle cyclic SCCs via fixpoint iteration.
-            // Process cyclic SCCs before the regular consistency group refresh
-            // so that SCC members are up-to-date when downstream STs check
-            // for upstream changes.
-            // Collect SCC member IDs so Step C can skip them.
-            let mut scc_member_ids: HashSet<i64> = HashSet::new();
-            if allow_circular {
-                let sccs = dag_ref.condensation_order();
-                for scc in &sccs {
-                    if !scc.is_cyclic {
-                        continue; // Singletons handled below in Step C
-                    }
-                    // Record members so Step C skips them.
-                    for node in &scc.nodes {
-                        if let NodeId::StreamTable(id) = node {
-                            scc_member_ids.insert(*id);
-                        }
-                    }
-                    // Check if any SCC member needs refresh
-                    let any_due = scc.nodes.iter().any(|node| {
-                        if let NodeId::StreamTable(id) = node {
-                            load_st_by_id(*id)
-                                .map(|st| {
-                                    (st.status == StStatus::Active
-                                        || st.status == StStatus::Initializing)
-                                        && (check_schedule(&st, dag_ref)
-                                            || check_upstream_changes(&st)
-                                            || st.needs_reinit)
-                                })
-                                .unwrap_or(false)
-                        } else {
-                            false
-                        }
-                    });
-                    if !any_due {
-                        continue;
-                    }
-                    iterate_to_fixpoint(
-                        scc,
-                        dag_ref,
-                        &mut retry_states,
-                        &retry_policy,
-                        now_ms,
-                        tick_watermark.as_deref(),
-                    );
-                }
-            }
-
-            // Step C: Compute consistency groups and refresh group-by-group
-            let groups = dag_ref.compute_consistency_groups();
-
-            for group in &groups {
-                let initial_table_changes: HashMap<i64, bool> = group
-                    .members
-                    .iter()
-                    .filter_map(|member| match member {
-                        NodeId::StreamTable(id) => {
-                            load_st_by_id(*id).map(|st| (*id, has_table_source_changes(&st)))
-                        }
-                        _ => None,
-                    })
-                    .collect();
-
-                if group.is_singleton() {
-                    // Fast path: no SAVEPOINT overhead for non-diamond STs.
-                    let pgt_id = match &group.members[0] {
-                        NodeId::StreamTable(id) => *id,
-                        _ => continue,
-                    };
-                    // Skip SCC members already handled by fixpoint iteration.
-                    if scc_member_ids.contains(&pgt_id) {
-                        continue;
-                    }
-                    // P3-5: Auto-backoff — skip this tick if the backoff
-                    // factor indicates we should wait longer.
-                    let bf = backoff_factors.get(&pgt_id).copied().unwrap_or(1.0);
-                    if bf > 1.0 {
-                        // Check if enough time has passed since last refresh
-                        // (effective interval = schedule * backoff_factor).
-                        // NOTE: We are already inside the outer BackgroundWorker::transaction,
-                        // so we must NOT call BackgroundWorker::transaction here — PostgreSQL
-                        // does not allow StartTransactionCommand when already in TBLOCK_STARTED
-                        // state (causes elog(FATAL)). Use Spi directly instead.
-                        let should_skip = if let Some(st) = load_st_by_id(pgt_id)
-                            && let Some(ref schedule_str) = st.schedule
-                            && let Ok(max_secs) = crate::api::parse_duration(schedule_str)
-                        {
-                            let effective_secs: f64 = max_secs as f64 * bf;
-                            let stale = Spi::get_one_with_args::<bool>(
-                                "SELECT CASE WHEN last_refresh_at IS NULL THEN true \
-                                 ELSE EXTRACT(EPOCH FROM (now() - last_refresh_at)) > $2 END \
-                                 FROM pgtrickle.pgt_stream_tables WHERE pgt_id = $1",
-                                &[st.pgt_id.into(), effective_secs.into()],
-                            )
-                            .unwrap_or(Some(false))
-                            .unwrap_or(false);
-                            !stale // skip if not yet due under backoff schedule
-                        } else {
-                            false
-                        };
-                        if should_skip {
-                            continue;
-                        }
-                    }
-                    let mut entry = drift_counters.entry(pgt_id).or_insert(0);
-                    refresh_single_st(
-                        pgt_id,
-                        dag_ref,
-                        now_ms,
-                        &mut retry_states,
-                        &retry_policy,
-                        initial_table_changes.get(&pgt_id).copied(),
-                        tick_watermark.as_deref(),
-                        Some(&mut entry),
-                        &mut spill_counters,
-                    );
-                    // P3-5: Update auto-backoff factor based on last refresh timing.
-                    // NOTE: We are already inside the outer BackgroundWorker::transaction,
-                    // so we must NOT call BackgroundWorker::transaction here — calling
-                    // StartTransactionCommand in TBLOCK_STARTED state causes elog(FATAL).
-                    if config::pg_trickle_auto_backoff() {
-                        update_backoff_factor(pgt_id, &mut backoff_factors);
-                    }
-                    continue;
-                }
-
-                // Multi-member group: check if all members have diamond_consistency = 'atomic'.
-                // Skip groups where all members are SCC-handled.
-                let non_scc_members: Vec<&NodeId> = group
-                    .members
-                    .iter()
-                    .filter(|m| {
-                        if let NodeId::StreamTable(id) = m {
-                            !scc_member_ids.contains(id)
-                        } else {
-                            true
-                        }
-                    })
-                    .collect();
-                if non_scc_members.is_empty() {
-                    continue;
-                }
-
-                let all_atomic = group.isolation_level
-                    == crate::dag::IsolationLevel::RepeatableRead
-                    || group.members.iter().all(|m| {
-                        if let NodeId::StreamTable(id) = m {
-                            load_st_by_id(*id)
-                                .map(|st| st.diamond_consistency == DiamondConsistency::Atomic)
-                                .unwrap_or(false)
-                        } else {
-                            false
-                        }
-                    });
-
-                if !all_atomic {
-                    // Not all members opted in — fall back to independent refreshes.
-                    for member in &group.members {
-                        let pgt_id = match member {
-                            NodeId::StreamTable(id) => *id,
-                            _ => continue,
-                        };
-                        let mut entry = drift_counters.entry(pgt_id).or_insert(0);
-                        refresh_single_st(
-                            pgt_id,
-                            dag_ref,
-                            now_ms,
-                            &mut retry_states,
-                            &retry_policy,
-                            initial_table_changes.get(&pgt_id).copied(),
-                            tick_watermark.as_deref(),
-                            Some(&mut entry),
-                            &mut spill_counters,
-                        );
-                    }
-                    continue;
-                }
-
-                // Atomic group: check group-level schedule policy.
-                let policy = group_schedule_policy(group);
-                if !is_group_due(group, policy, dag_ref) {
-                    // When the Slowest policy is active, the group waits for all
-                    // members to be due simultaneously.  In asymmetric pipelines
-                    // (e.g. a diamond where only one branch has CDC events on a
-                    // given tick), convergence nodes can accumulate staleness
-                    // indefinitely.  Emit stale-data alerts so operators are
-                    // notified rather than silently waiting.
-                    if policy == DiamondSchedulePolicy::Slowest {
-                        for node in &group.convergence_points {
-                            if let NodeId::StreamTable(id) = node
-                                && let Some(conv_st) = load_st_by_id(*id)
-                            {
-                                emit_stale_alert_if_needed(&conv_st);
-                            }
-                        }
-                    }
-                    continue;
-                }
-
-                // All members due (per policy) — wrap in an internal sub-transaction.
-                //
-                // Background workers cannot use `Spi::run("SAVEPOINT …")` because
-                // PostgreSQL rejects transaction-control commands issued via SPI
-                // (SPI_ERROR_TRANSACTION).  The correct approach is the internal
-                // C-level sub-transaction API which bypasses SPI entirely.
-                //
-                // SAFETY: BeginInternalSubTransaction sets up a sub-transaction
-                // within the current worker transaction using PostgreSQL's resource-
-                // owner mechanism.  We save and restore CurrentMemoryContext and
-                // CurrentResourceOwner to ensure no context leak regardless of
-                // whether the sub-transaction commits or rolls back.  All work
-                // inside is done through SPI (which handles its own C-level exception
-                // boundaries), so a Rust-visible panic from this block is not
-                // expected.
-                let subtxn = SubTransaction::begin();
-
-                let mut group_ok = true;
-                let mut refreshed_ids: Vec<i64> = Vec::new();
-
-                for member in &group.members {
-                    let pgt_id = match member {
-                        NodeId::StreamTable(id) => *id,
-                        _ => continue,
-                    };
-
-                    let st = match load_st_by_id(pgt_id) {
-                        Some(st) => st,
-                        None => continue,
-                    };
-
-                    // Skip non-active STs
-                    if st.status != StStatus::Active && st.status != StStatus::Initializing {
-                        continue;
-                    }
-
-                    // BOOT-4: Skip if any source is gated.
-                    let gated_oids = load_gated_source_oids();
-                    if is_any_source_gated(pgt_id, &gated_oids) {
-                        log!(
-                            "pg_trickle: skipping {}.{} in atomic group — source gated",
-                            st.pgt_schema,
-                            st.pgt_name,
-                        );
-                        log_gated_skip(&st);
-                        continue;
-                    }
-
-                    // WM-4: Skip if watermarks misaligned.
-                    let (wm_misaligned, wm_reason) = is_watermark_misaligned(pgt_id);
-                    if wm_misaligned {
-                        let reason = wm_reason.as_deref().unwrap_or("watermark misaligned");
-                        log!(
-                            "pg_trickle: skipping {}.{} in atomic group — {}",
-                            st.pgt_schema,
-                            st.pgt_name,
-                            reason,
-                        );
-                        log_watermark_skip(&st, reason);
-                        continue;
-                    }
-
-                    // WM-7: Skip if any source watermark is stuck.
-                    let (wm_stuck, wm_stuck_reason) = is_watermark_stuck(pgt_id);
-                    if wm_stuck {
-                        let reason = wm_stuck_reason.as_deref().unwrap_or("watermark stuck");
-                        log!(
-                            "pg_trickle: skipping {}.{} in atomic group — {}",
-                            st.pgt_schema,
-                            st.pgt_name,
-                            reason,
-                        );
-                        log_watermark_skip(&st, reason);
-                        continue;
-                    }
-
-                    // Check retry backoff
-                    let retry = retry_states.entry(pgt_id).or_default();
-                    if retry.is_in_backoff(now_ms) {
-                        emit_stale_alert_if_needed(&st);
-                        continue;
-                    }
-
-                    // Skip if catalog row locked by another session
-                    if check_skip_needed(&st) {
-                        continue;
-                    }
-
-                    // FUSE-5: Check fuse circuit breaker.
-                    if evaluate_fuse(&st) {
-                        log!(
-                            "pg_trickle: {}.{} fuse blown — skipping in diamond group",
-                            st.pgt_schema,
-                            st.pgt_name,
-                        );
-                        continue;
-                    }
-
-                    let (has_changes, _has_stream_table_changes) =
-                        upstream_change_state(&st, initial_table_changes.get(&pgt_id).copied());
-                    let action = refresh::determine_refresh_action(&st, has_changes);
-                    let result =
-                        execute_scheduled_refresh(&st, action, tick_watermark.as_deref(), None);
-
-                    match result {
-                        RefreshOutcome::Success => {
-                            refreshed_ids.push(pgt_id);
-                        }
-                        RefreshOutcome::RetryableFailure | RefreshOutcome::PermanentFailure => {
-                            log!(
-                                "pg_trickle: diamond group rollback — member {}.{} failed",
-                                st.pgt_schema,
-                                st.pgt_name,
-                            );
-                            group_ok = false;
-                            break;
-                        }
-                    }
-                }
-
-                if group_ok {
-                    subtxn.commit();
-                    // Reset retry states for all refreshed members
-                    for id in &refreshed_ids {
-                        retry_states.entry(*id).or_default().reset();
-                    }
-                } else {
-                    subtxn.rollback();
-                    // Record failure for retry tracking on all attempted members
-                    for id in &refreshed_ids {
-                        let retry = retry_states.entry(*id).or_default();
-                        retry.record_failure(&retry_policy, now_ms);
-                    }
-                    log!(
-                        "pg_trickle: diamond group rolled back ({} members)",
-                        group.members.len(),
-                    );
-                }
-            }
-
-            // Step D & E: Handled in the pre-refresh transaction above.
-
-            // Step F: Prune retry states for STs that no longer exist
-            // (avoid accumulating stale state)
-            let active_ids: std::collections::HashSet<i64> = groups
-                .iter()
-                .flat_map(|g| g.members.iter())
-                .filter_map(|n| match n {
-                    NodeId::StreamTable(id) => Some(*id),
-                    _ => None,
-                })
-                .collect();
-            retry_states.retain(|id, _| active_ids.contains(id));
-        }));
-
-        // Phase 4: Spawn workers outside the transaction (after commit).
-        for (db, job_id) in pending_spawns.drain(..) {
-            if let Err(e) = spawn_refresh_worker(&db, job_id) {
-                shmem::release_worker_token();
-                parallel_state.per_db_inflight = parallel_state.per_db_inflight.saturating_sub(1);
-                // Find and clear the in-flight tracking for this job.
-                for us in parallel_state.unit_states.values_mut() {
-                    if us.inflight_job_id == Some(job_id) {
-                        us.inflight_job_id = None;
-                        break;
-                    }
-                }
-                log!(
-                    "pg_trickle: parallel dispatch — failed to spawn worker for job {}: {}",
-                    job_id,
-                    e,
-                );
-                // Cancel the enqueued job row.
-                BackgroundWorker::transaction(AssertUnwindSafe(|| {
-                    let _ = SchedulerJob::cancel(job_id, &format!("Worker spawn failed: {}", e));
-                }));
-            }
-        }
-    }
 }
 
 // ── Refresh Outcome ────────────────────────────────────────────────────────
@@ -6551,7 +3800,13 @@ fn compute_freshness_deadline(st: &StreamTableMeta) -> Option<TimestampWithTimeZ
 
 #[cfg(test)]
 mod tests {
+    use super::dispatch::{
+        ADAPTIVE_POLL_MIN_MS, ParallelDispatchState, UnitDispatchState, compute_adaptive_poll_ms,
+        parse_worker_extra, sort_ready_queue_by_priority,
+    };
     use super::*;
+    use crate::dag::{ExecutionUnitDag, ExecutionUnitId};
+    use std::collections::VecDeque;
 
     #[test]
     fn test_current_epoch_ms_returns_reasonable_value() {
